Fix #77 - Allow `configURL` to be passed for custom types
This MR basically allows already parsed configs to work with the proper
relative or absolute URL within fetch, files, or packages operations.

  * added a `configURL` field that overload the `config` when this is an object
  * still fallback to a local `./config.txt` as it was doing before for backward compatibility
  * remove the circular dependency around io and stdio
  * resolve without breaking Worker callbacks as it's done in PyScript, exposing the same library too
  * added and tested a use case that used to fail when config as object was already passed and its files operations were breaking
