+++ docs/core.js
@@ -1,3 +1,3 @@
-const e=(e,t=document)=>[...t.querySelectorAll(e)],t=(e,t=document)=>{const n=(new XPathEvaluator).createExpression(e).evaluate(t,XPathResult.ORDERED_NODE_SNAPSHOT_TYPE),r=[];for(let e=0,{snapshotLength:t}=n;e<t;e++)r.push(n.snapshotItem(e));return r},n="object"==typeof self?self:globalThis,r=e=>((e,t)=>{const r=(t,n)=>(e.set(n,t),t),s=o=>{if(e.has(o))return e.get(o);const[a,i]=t[o];switch(a){case 0:case-1:return r(i,o);case 1:{const e=r([],o);for(const t of i)e.push(s(t));return e}case 2:{const e=r({},o);for(const[t,n]of i)e[s(t)]=s(n);return e}case 3:return r(new Date(i),o);case 4:{const{source:e,flags:t}=i;return r(new RegExp(e,t),o)}case 5:{const e=r(new Map,o);for(const[t,n]of i)e.set(s(t),s(n));return e}case 6:{const e=r(new Set,o);for(const t of i)e.add(s(t));return e}case 7:{const{name:e,message:t}=i;return r(new n[e](t),o)}case 8:return r(BigInt(i),o);case"BigInt":return r(Object(BigInt(i)),o)}return r(new n[a](i),o)};return s})(new Map,e)(0),s="",{toString:o}={},{keys:a}=Object,i=e=>{const t=typeof e;if("object"!==t||!e)return[0,t];const n=o.call(e).slice(8,-1);switch(n){case"Array":return[1,s];case"Object":return[2,s];case"Date":return[3,s];case"RegExp":return[4,s];case"Map":return[5,s];case"Set":return[6,s]}return n.includes("Array")?[1,n]:n.includes("Error")?[7,n]:[2,n]},c=([e,t])=>0===e&&("function"===t||"symbol"===t),l=(e,{json:t,lossy:n}={})=>{const r=[];return((e,t,n,r)=>{const s=(e,t)=>{const s=r.push(e)-1;return n.set(t,s),s},o=r=>{if(n.has(r))return n.get(r);let[l,u]=i(r);switch(l){case 0:{let t=r;switch(u){case"bigint":l=8,t=r.toString();break;case"function":case"symbol":if(e)throw new TypeError("unable to serialize "+u);t=null;break;case"undefined":return s([-1],r)}return s([l,t],r)}case 1:{if(u)return s([u,[...r]],r);const e=[],t=s([l,e],r);for(const t of r)e.push(o(t));return t}case 2:{if(u)switch(u){case"BigInt":return s([u,r.toString()],r);case"Boolean":case"Number":case"String":return s([u,r.valueOf()],r)}if(t&&"toJSON"in r)return o(r.toJSON());const n=[],p=s([l,n],r);for(const t of a(r))!e&&c(i(r[t]))||n.push([o(t),o(r[t])]);return p}case 3:return s([l,r.toISOString()],r);case 4:{const{source:e,flags:t}=r;return s([l,{source:e,flags:t}],r)}case 5:{const t=[],n=s([l,t],r);for(const[n,s]of r)(e||!c(i(n))&&!c(i(s)))&&t.push([o(n),o(s)]);return n}case 6:{const t=[],n=s([l,t],r);for(const n of r)!e&&c(i(n))||t.push(o(n));return n}}const{message:p}=r;return s([l,{name:u,message:p}],r)};return o})(!(t||n),!!t,new Map,r)(e),r},{parse:u,stringify:p}=JSON,f={json:!0,lossy:!0};var h=Object.freeze({__proto__:null,parse:e=>r(u(e)),stringify:e=>p(l(e,f))});const d="ac6c1995-0675-46af-a99f-7fa298fa4e6d",g="M"+d,y="T"+d,m="array",w="function",_="null",b="number",E="object",v="string",k="symbol",S="undefined",x="apply",T="construct",A="defineProperty",O="deleteProperty",j="get",R="getOwnPropertyDescriptor",$="getPrototypeOf",P="has",N="isExtensible",I="ownKeys",M="preventExtensions",F="set",H="setPrototypeOf",W="delete";var C=e=>({value:new Promise((t=>{let n=new Worker("data:application/javascript,onmessage%3D(%7Bdata%3Ab%7D)%3D%3E(Atomics.wait(b%2C0)%2CpostMessage(0))");n.onmessage=t,n.postMessage(e)}))})
-/*! (c) Andrea Giammarchi - ISC */;const{Int32Array:L,Map:D,SharedArrayBuffer:J,Uint16Array:B}=globalThis,{BYTES_PER_ELEMENT:U}=L,{BYTES_PER_ELEMENT:q}=B,{isArray:G}=Array,{notify:K,wait:X,waitAsync:Y}=Atomics,z=new WeakSet,V=new WeakMap,Z={value:{then:e=>e()}};let Q=0;const ee=(e,{parse:t=JSON.parse,stringify:n=JSON.stringify,transform:r,interrupt:s}=JSON)=>{if(!V.has(e)){const o=(t,...n)=>e.postMessage({[d]:n},{transfer:t}),a=typeof s===w?s:s?.handler,i=s?.delay||42,c=new TextDecoder("utf-16"),l=(e,t)=>e?(Y||C)(t,0):(a?((e,t,n)=>{for(;"timed-out"===X(e,0,0,t);)n()})(t,i,a):X(t,0),Z);let u=!1;V.set(e,new Proxy(new D,{[P]:(e,t)=>"string"==typeof t&&!t.startsWith("_"),[j]:(n,s)=>"then"===s?null:(...n)=>{const a=Q++;let i=new L(new J(2*U)),p=[];z.has(n.at(-1)||p)&&z.delete(p=n.pop()),o(p,a,i,s,r?n.map(r):n);const f=e!==globalThis;let h=0;return u&&f&&(h=setTimeout(console.warn,1e3,`💀🔒 - Possible deadlock if proxy.${s}(...args) is awaited`)),l(f,i).value.then((()=>{clearTimeout(h);const e=i[1];if(!e)return;const n=q*e;return i=new L(new J(n+n%U)),o([],a,i),l(f,i).value.then((()=>t(c.decode(new B(i.buffer).slice(0,e)))))}))},[F](t,s,o){const a=typeof o;if(a!==w)throw new Error(`Unable to assign ${s} as ${a}`);if(!t.size){const s=new D;e.addEventListener("message",(async e=>{const o=e.data?.[d];if(G(o)){e.stopImmediatePropagation();const[a,i,...c]=o;let l;if(c.length){const[e,o]=c;if(t.has(e)){u=!0;try{const c=await t.get(e)(...o);if(void 0!==c){const e=n(r?r(c):c);s.set(a,e),i[1]=e.length}}catch(e){l=e}finally{u=!1}}else l=new Error(`Unsupported action: ${e}`);i[0]=1}else{const e=s.get(a);s.delete(a);for(let t=new B(i.buffer),n=0;n<e.length;n++)t[n]=e.charCodeAt(n)}if(K(i,0),l)throw l}}))}return!!t.set(s,o)}}))}return V.get(e)};ee.transfer=(...e)=>(z.add(e),e);const{isArray:te}=Array,ne=(e,t)=>t,re=e=>typeof e===w?(e=>e())(e):e;function se(){return this}const oe=(e,t)=>e===m?[t]:{t:e,v:t},ae=(e,t=ne)=>{let n=typeof e,r=e;return n===E&&(te(e)?(n=m,r=e.at(0)):({t:n,v:r}=e)),t(n,r)},ie=(e,t)=>e===w?t:oe(e,t),ce=(e,t=ie)=>{const n=null===e?_:typeof e;return t(n===E&&te(e)?m:n,e)},le=new FinalizationRegistry((([e,t,n])=>{n&&console.debug(`Held value ${String(t)} not relevant anymore`),e(t)})),ue=Object.create(null),pe=(e,t,{debug:n,return:r,token:s=e}=ue)=>{const o=r||new Proxy(e,ue),a=[o,[t,e,!!n]];return!1!==s&&a.push(s),le.register(...a),o},{defineProperty:fe,deleteProperty:he,getOwnPropertyDescriptor:de,getPrototypeOf:ge,isExtensible:ye,ownKeys:me,preventExtensions:we,set:_e,setPrototypeOf:be}=Reflect,{assign:Ee,create:ve}=Object,ke=ge(Int8Array),Se=(e,t)=>{const{get:n,set:r,value:s}=e;return n&&(e.get=t(n)),r&&(e.set=t(r)),s&&(e.value=t(s)),e},xe=e=>t=>ce(t,((t,n)=>{switch(t){case _:return oe(_,n);case E:if(n===globalThis)return oe(t,null);case m:case w:return e(t,n);case"boolean":case b:case v:case S:case"bigint":return oe(t,n);case k:{if(Te.has(n))return oe(t,Te.get(n));let e=Symbol.keyFor(n);if(e)return oe(t,`.${e}`)}}throw new TypeError(`Unable to handle this ${t}: ${String(n)}`)})),Te=new Map(me(Symbol).filter((e=>typeof Symbol[e]===k)).map((e=>[Symbol[e],e]))),Ae=e=>{if(e.startsWith("."))return Symbol.for(e.slice(1));for(const[t,n]of Te)if(n===e)return t},Oe=e=>e;var je=((e,t)=>{const n=t&&new WeakMap;if(t){const{addEventListener:e}=EventTarget.prototype;fe(EventTarget.prototype,"addEventListener",{value(t,r,...s){return s.at(0)?.invoke&&(n.has(this)||n.set(this,new Map),n.get(this).set(t,[].concat(s[0].invoke)),delete s[0].invoke),e.call(this,t,r,...s)}})}const r=t&&(e=>{const{currentTarget:t,target:r,type:s}=e;for(const o of n.get(t||r)?.get(s)||[])e[o]()});return function(n,s,o,...a){let i=0,c=this?.transform||Oe;const l=new Map,u=new Map,{[o]:p}=n,f=a.length?Ee(ve(globalThis),...a):globalThis,h=xe(((e,t)=>{if(!l.has(t)){let n;for(;u.has(n=i++););l.set(t,n),u.set(n,e===w?t:c(t))}return oe(e,l.get(t))})),d=e=>{p(W,oe(v,e))},g=(e,n)=>{switch(e){case E:if(null==n)return f;case m:if(typeof n===b)return u.get(n);if(!(n instanceof ke))for(const e in n)n[e]=y(n[e]);return n;case w:if(typeof n===v){if(!u.has(n)){const e=function(...e){return t&&e.at(0)instanceof Event&&r(...e),p(x,oe(w,n),h(this),e.map(h))};return u.set(n,new WeakRef(e)),pe(n,d,{return:e,token:!1})}return u.get(n).deref()}return u.get(n);case k:return Ae(n)}return n},y=e=>ae(e,g),_={[x]:(e,t,n)=>h(e.apply(t,n)),[T]:(e,t)=>h(new e(...t)),[A]:(e,t,n)=>h(fe(e,t,n)),[O]:(e,t)=>h(he(e,t)),[$]:e=>h(ge(e)),[j]:(e,t)=>h(e[t]),[R]:(e,t)=>{const n=de(e,t);return n?oe(E,Se(n,h)):oe(S,n)},[P]:(e,t)=>h(t in e),[N]:e=>h(ye(e)),[I]:e=>oe(m,me(e).map(h)),[M]:e=>h(we(e)),[F]:(e,t,n)=>h(_e(e,t,n)),[H]:(e,t)=>h(be(e,t)),[W](e){l.delete(u.get(e)),u.delete(e)}};return n[s]=(e,t,...n)=>{switch(e){case x:n[0]=y(n[0]),n[1]=n[1].map(y);break;case T:n[0]=n[0].map(y);break;case A:{const[e,t]=n;n[0]=y(e);const{get:r,set:s,value:o}=t;r&&(t.get=y(r)),s&&(t.set=y(s)),o&&(t.value=y(o));break}default:n=n.map(y)}return _[e](y(t),...n)},{proxy:n,[e.toLowerCase()]:f,[`is${e}Proxy`]:()=>!1}}})("Window",!0),Re=(e=>{let t=0;const n=new Map,r=new Map,s=Symbol();return function(o,a,i){const c=this?.transform||Oe,{[a]:l}=o,u=new Map,p=e=>{u.delete(e),l(W,f(e))},f=xe(((e,o)=>{if(s in o)return re(o[s]);if(e===w){if(o=c(o),!r.has(o)){let e;for(;r.has(e=String(t++)););n.set(o,e),r.set(e,o)}return oe(e,n.get(o))}if(!(o instanceof ke)){o=c(o);for(const e in o)o[e]=f(o[e])}return oe(e,o)})),h=(e,t,n)=>{if(!u.has(n)){const r=t===w?(e=>se.bind(e))(e):e,s=new Proxy(r,y);return u.set(n,new WeakRef(s)),pe(n,p,{return:s,token:!1})}return u.get(n).deref()},d=e=>ae(e,((t,n)=>{switch(t){case E:if(null===n)return globalThis;case m:return typeof n===b?h(e,t,n):n;case w:return typeof n===v?r.get(n):h(e,t,n);case k:return Ae(n)}return n})),g=(e,t,...n)=>d(l(e,re(t),...n)),y={[x]:(e,t,n)=>g(x,e,f(t),n.map(f)),[T]:(e,t)=>g(T,e,t.map(f)),[A]:(e,t,n)=>{const{get:r,set:s,value:o}=n;return typeof r===w&&(n.get=f(r)),typeof s===w&&(n.set=f(s)),typeof o===w&&(n.value=f(o)),g(A,e,f(t),n)},[O]:(e,t)=>g(O,e,f(t)),[$]:e=>g($,e),[j]:(e,t)=>t===s?e:g(j,e,f(t)),[R]:(e,t)=>{const n=g(R,e,f(t));return n&&Se(n,d)},[P]:(e,t)=>t===s||g(P,e,f(t)),[N]:e=>g(N,e),[I]:e=>g(I,e).map(d),[M]:e=>g(M,e),[F]:(e,t,n)=>g(F,e,f(t),f(n)),[H]:(e,t)=>g(H,e,f(t))};o[i]=(e,t,s,o)=>{switch(e){case x:return d(t).apply(d(s),o.map(d));case W:{const e=d(t);n.delete(r.get(e)),r.delete(e)}}};const _=new Proxy(oe(E,null),y);return{[e.toLowerCase()]:_,[`is${e}Proxy`]:e=>typeof e===E&&!!e&&s in e,proxy:o}}})("Window"),$e=typeof Worker===w?Worker:class{};const Pe=new WeakMap,Ne=(e,...t)=>{const n=ee(e,...t);if(!Pe.has(n)){const r=e instanceof $e?je:Re;Pe.set(n,r.call(t.at(0),n,g,y))}return Pe.get(n)};Ne.transfer=ee.transfer;Promise.withResolvers||(Promise.withResolvers=function(){var e,t,n=new this((function(n,r){e=n,t=r}));return{resolve:e,reject:t,promise:n}});const Ie=e=>e.arrayBuffer(),Me=e=>e.json(),Fe=e=>e.text();const He={object(...e){return this.string(function(e){for(var t=e[0],n=1,r=arguments.length;n<r;n++)t+=arguments[n]+e[n];return t}(...e))},string(e){for(const t of e.split(/[\r\n]+/))if(t.trim().length){/^(\s+)/.test(t)&&(e=e.replace(new RegExp("^"+RegExp.$1,"gm"),""));break}return e}},{replace:We}="",Ce=/&(?:amp|#38|lt|#60|gt|#62|apos|#39|quot|#34);/g,Le={"&amp;":"&","&#38;":"&","&lt;":"<","&#60;":"<","&gt;":">","&#62;":">","&apos;":"'","&#39;":"'","&quot;":'"',"&#34;":'"'},De=e=>Le[e],Je=(e,...t)=>He[typeof e](e,...t),Be=e=>We.call(e,Ce,De),{isArray:Ue}=Array,{assign:qe,create:Ge,defineProperties:Ke,defineProperty:Xe,entries:Ye}=Object,{all:ze,resolve:Ve}=new Proxy(Promise,{get:(e,t)=>e[t].bind(e)}),Ze=(e,t=location.href)=>new URL(e,t.replace(/^blob:/,"")).href;let Qe=0;const et=(e,t)=>({id:e.id||(e.id=`${t}-w${Qe++}`),tag:e.tagName}),tt=(e,t,n,r=!1,s=CustomEvent)=>{e.dispatchEvent(new s(`${t}:${n}`,{bubbles:!0,detail:{worker:r}}))},nt=(e,t,n,r)=>({type:t,config:n,interpreter:r,io:ht.get(r),run:(t,...n)=>e.run(r,t,...n),runAsync:(t,...n)=>e.runAsync(r,t,...n),runEvent:(...t)=>e.runEvent(r,...t)}),rt=(e,t)=>(n,r)=>{const s=e[t].bind(e);e[t]=(e,t,...o)=>s(e,`${r?n:t}\n${r?t:n}`,...o)},st=Symbol.for("polyscript.js_modules"),ot=new Map;Xe(globalThis,st,{value:ot});const at=new Proxy(ot,{get:(e,t)=>e.get(t),has:(e,t)=>e.has(t),ownKeys:e=>[...e.keys()]}),it=(e,t)=>!t.startsWith("_"),ct=(e,t)=>new Proxy(e,{has:it,get:(e,n)=>e[t][n]}),lt=(e,t,n,r)=>{if("pyodide"===e)return;const s="polyscript.js_modules";for(const e of Reflect.ownKeys(r))t.registerJSModule(n,`${s}.${e}`,ct(r,e));t.registerJSModule(n,s,r)},ut=(e,t)=>import(e).then((e=>{ot.set(t,{...e})})),pt=e=>new Promise(((t,n)=>{document.querySelector(`link[href="${e}"]`)&&t(),document.head.append(qe(document.createElement("link"),{rel:"stylesheet",href:e,onload:t,onerror:n}))})),ft=e=>/\.css/i.test(new URL(e).pathname),ht=new WeakMap,dt=e=>{const t=e||console,n={stderr:(t.stderr||console.error).bind(t),stdout:(t.stdout||console.log).bind(t)};return{stderr:(...e)=>n.stderr(...e),stdout:(...e)=>n.stdout(...e),async get(e){const t=await e;return ht.set(t,n),t}}},gt=({FS:e,PATH:t,PATH_FS:n},r,s)=>{const o=n.resolve(r);return e.mkdirTree(t.dirname(o)),e.writeFile(o,new Uint8Array(s),{canOwn:!0})},yt=e=>{const t=e.split("/");return t.pop(),t.join("/")},mt=(e,t)=>{const n=[];for(const r of t.split("/"))"."!==r&&(n.push(r),r&&e.mkdir(n.join("/")))},wt=(e,t)=>{const n=[];for(const e of t.split("/"))switch(e){case"":case".":break;case"..":n.pop();break;default:n.push(e)}return[e.cwd()].concat(n).join("/").replace(/^\/+/,"/")},_t=e=>{const t=e.map((e=>e.trim().replace(/(^[/]*|[/]*$)/g,""))).filter((e=>""!==e&&"."!==e)).join("/");return e[0].startsWith("/")?`/${t}`:t},bt=(e,t)=>fetch(Ze(t,Et.get(e))),Et=new WeakMap,vt=(e,t,n)=>ze((e=>{for(const{files:t,to_file:n,from:r=""}of e){if(void 0!==t&&void 0!==n)throw new Error("Cannot use 'to_file' and 'files' parameters together!");if(void 0===t&&void 0===n&&r.endsWith("/"))throw new Error(`Couldn't determine the filename from the path ${r}, please supply 'to_file' parameter.`)}return e.flatMap((({from:e="",to_folder:t=".",to_file:n,files:r})=>{if(Ue(r))return r.map((n=>({url:_t([e,n]),path:_t([t,n])})));const s=n||e.slice(1+e.lastIndexOf("/"));return[{url:e,path:_t([t,s])}]}))})(n).map((({url:r,path:s})=>bt(n,r).then(Ie).then((n=>e.writeFile(t,s,n)))))),kt=(e,t)=>t.endsWith("/")?`${t}${e.split("/").pop()}`:t,St=(e,t)=>e.replace(/\{.+?\}/g,(e=>{if(!t.has(e))throw new SyntaxError(`Invalid template: ${e}`);return t.get(e)})),xt=(e,t,n)=>ze((e=>{const t=new Map,n=new Set,r=[];for(const[s,o]of Ye(e))if(/^\{.+\}$/.test(s)){if(t.has(s))throw new SyntaxError(`Duplicated template: ${s}`);t.set(s,St(o,t))}else{const e=St(s,t),a=kt(e,St(o||"./",t));if(n.has(a))throw new SyntaxError(`Duplicated destination: ${a}`);n.add(a),r.push({url:e,path:a})}return r})(n).map((({url:r,path:s})=>bt(n,r).then(Ie).then((n=>e.writeFile(t,s,n)))))),Tt=typeof document===S,At=({main:e,worker:t})=>{const n=[];if(t&&Tt)for(let[e,r]of Ye(t))e=Ze(e,Et.get(t)),n.push(ut(e,r));if(e&&!Tt)for(let[t,r]of Ye(e))t=Ze(t,Et.get(e)),ft(t)?pt(t):n.push(ut(t,r));return ze(n)},Ot=(e,t,n)=>{e.registerJsModule(t,n)},jt=(e,t,...n)=>{try{return e.runPython(Je(t),...n)}catch(t){ht.get(e).stderr(t)}},Rt=async(e,t,...n)=>{try{return await e.runPythonAsync(Je(t),...n)}catch(t){ht.get(e).stderr(t)}},$t=async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.globals.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){ht.get(e).stderr(t)}};var Pt=(new TextEncoder).encode('from uio import StringIO\nimport sys\n\nclass Response:\n    def __init__(self, f):\n        self.raw = f\n        self.encoding = "utf-8"\n        self._cached = None\n\n    def close(self):\n        if self.raw:\n            self.raw.close()\n            self.raw = None\n        self._cached = None\n\n    @property\n    def content(self):\n        if self._cached is None:\n            try:\n                self._cached = self.raw.read()\n            finally:\n                self.raw.close()\n                self.raw = None\n        return self._cached\n\n    @property\n    def text(self):\n        return str(self.content, self.encoding)\n\n    def json(self):\n        import ujson\n\n        return ujson.loads(self.content)\n\n\n# TODO try to support streaming xhr requests, a-la pyodide-http\nHEADERS_TO_IGNORE = ("user-agent",)\n\n\ntry:\n    import js\nexcept Exception as err:\n    raise OSError("This version of urequests can only be used in the browser")\n\n# TODO try to support streaming xhr requests, a-la pyodide-http\n\nHEADERS_TO_IGNORE = ("user-agent",)\n\n\ndef request(\n    method,\n    url,\n    data=None,\n    json=None,\n    headers={},\n    stream=None,\n    auth=None,\n    timeout=None,\n    parse_headers=True,\n):\n    from js import XMLHttpRequest\n\n    xhr = XMLHttpRequest.new()\n    xhr.withCredentials = False\n\n    if auth is not None:\n        import ubinascii\n\n        username, password = auth\n        xhr.open(method, url, False, username, password)\n    else:\n        xhr.open(method, url, False)\n\n    for name, value in headers.items():\n        if name.lower() not in HEADERS_TO_IGNORE:\n            xhr.setRequestHeader(name, value)\n\n    if timeout:\n        xhr.timeout = int(timeout * 1000)\n\n    if json is not None:\n        assert data is None\n        import ujson\n\n        data = ujson.dumps(json)\n        # s.write(b"Content-Type: application/json\\r\\n")\n        xhr.setRequestHeader("Content-Type", "application/json")\n\n    xhr.send(data)\n\n    # Emulates the construction process in the original urequests\n    resp = Response(StringIO(xhr.responseText))\n    resp.status_code = xhr.status\n    resp.reason = xhr.statusText\n    resp.headers = xhr.getAllResponseHeaders()\n\n    return resp\n\n\n# Other methods - head, post, put, patch, delete - are not used by\n# mip and therefore not included\n\n\ndef get(url, **kw):\n    return request("GET", url, **kw)\n\n\n# Content below this line is from the Micropython MIP package and is covered\n# by the applicable MIT license:\n# \n# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\n# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \n# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\n# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \n# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \n# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER \n# DEALINGS IN THE SOFTWARE.\n\n# MicroPython package installer\n# MIT license; Copyright (c) 2022 Jim Mussared\n\n\n_PACKAGE_INDEX = const("https://micropython.org/pi/v2")\n_CHUNK_SIZE = 128\n\n\n# This implements os.makedirs(os.dirname(path))\ndef _ensure_path_exists(path):\n    import os\n\n    split = path.split("/")\n\n    # Handle paths starting with "/".\n    if not split[0]:\n        split.pop(0)\n        split[0] = "/" + split[0]\n\n    prefix = ""\n    for i in range(len(split) - 1):\n        prefix += split[i]\n        try:\n            os.stat(prefix)\n        except:\n            os.mkdir(prefix)\n        prefix += "/"\n\n\n# Copy from src (stream) to dest (function-taking-bytes)\ndef _chunk(src, dest):\n    buf = memoryview(bytearray(_CHUNK_SIZE))\n    while True:\n        n = src.readinto(buf)\n        if n == 0:\n            break\n        dest(buf if n == _CHUNK_SIZE else buf[:n])\n\n\n# Check if the specified path exists and matches the hash.\ndef _check_exists(path, short_hash):\n    import os\n\n    try:\n        import binascii\n        import hashlib\n\n        with open(path, "rb") as f:\n            hs256 = hashlib.sha256()\n            _chunk(f, hs256.update)\n            existing_hash = str(binascii.hexlify(hs256.digest())[: len(short_hash)], "utf-8")\n            return existing_hash == short_hash\n    except:\n        return False\n\n\ndef _rewrite_url(url, branch=None):\n    if not branch:\n        branch = "HEAD"\n    if url.startswith("github:"):\n        url = url[7:].split("/")\n        url = (\n            "https://raw.githubusercontent.com/"\n            + url[0]\n            + "/"\n            + url[1]\n            + "/"\n            + branch\n            + "/"\n            + "/".join(url[2:])\n        )\n    return url\n\n\ndef _download_file(url, dest):\n    response = get(url)\n    try:\n        if response.status_code != 200:\n            print("Error", response.status_code, "requesting", url)\n            return False\n\n        print("Copying:", dest)\n        _ensure_path_exists(dest)\n        with open(dest, "wb") as f:\n            _chunk(response.raw, f.write)\n\n        return True\n    finally:\n        response.close()\n\n\ndef _install_json(package_json_url, index, target, version, mpy):\n    response = get(_rewrite_url(package_json_url, version))\n    try:\n        if response.status_code != 200:\n            print("Package not found:", package_json_url)\n            return False\n\n        package_json = response.json()\n    finally:\n        response.close()\n    for target_path, short_hash in package_json.get("hashes", ()):\n        fs_target_path = target + "/" + target_path\n        if _check_exists(fs_target_path, short_hash):\n            print("Exists:", fs_target_path)\n        else:\n            file_url = "{}/file/{}/{}".format(index, short_hash[:2], short_hash)\n            if not _download_file(file_url, fs_target_path):\n                print("File not found: {} {}".format(target_path, short_hash))\n                return False\n    for target_path, url in package_json.get("urls", ()):\n        fs_target_path = target + "/" + target_path\n        if not _download_file(_rewrite_url(url, version), fs_target_path):\n            print("File not found: {} {}".format(target_path, url))\n            return False\n    for dep, dep_version in package_json.get("deps", ()):\n        if not _install_package(dep, index, target, dep_version, mpy):\n            return False\n    return True\n\n\ndef _install_package(package, index, target, version, mpy):\n    if (\n        package.startswith("http://")\n        or package.startswith("https://")\n        or package.startswith("github:")\n    ):\n        if package.endswith(".py") or package.endswith(".mpy"):\n            print("Downloading {} to {}".format(package, target))\n            return _download_file(\n                _rewrite_url(package, version), target + "/" + package.rsplit("/")[-1]\n            )\n        else:\n            if not package.endswith(".json"):\n                if not package.endswith("/"):\n                    package += "/"\n                package += "package.json"\n            print("Installing {} to {}".format(package, target))\n    else:\n        if not version:\n            version = "latest"\n        print("Installing {} ({}) from {} to {}".format(package, version, index, target))\n\n        mpy_version = (\n            sys.implementation._mpy & 0xFF if mpy and hasattr(sys.implementation, "_mpy") else "py"\n        )\n\n        package = "{}/package/{}/{}/{}.json".format(index, mpy_version, package, version)\n\n    return _install_json(package, index, target, version, mpy)\n\n\ndef install(package, index=None, target=None, version=None, mpy=True):\n    if not target:\n        for p in sys.path:\n            if p.endswith("/lib"):\n                target = p\n                break\n        else:\n            print("Unable to find lib dir in sys.path")\n            return\n\n    if not index:\n        index = _PACKAGE_INDEX\n\n    if _install_package(package, index.rstrip("/"), target, version, mpy):\n        print("Done")\n    else:\n        print("Package may be partially installed")\n');var Nt={type:"micropython",module:(e="1.21.0-279")=>`https://cdn.jsdelivr.net/npm/@micropython/micropython-webassembly-pyscript@${e}/micropython.mjs`,async engine({loadMicroPython:e},t,n){const{stderr:r,stdout:s,get:o}=dt();n=n.replace(/\.m?js$/,".wasm");const a=await o(e({stderr:r,stdout:s,url:n}));if(t.files&&await xt(this,a,t.files),t.fetch&&await vt(this,a,t.fetch),t.js_modules&&await At(t.js_modules),this.writeFile(a,"./mip.py",Pt),t.packages){const e=a.pyimport("mip");for(const n of t.packages)e.install(n)}return a},registerJSModule:Ot,run:jt,runAsync:Rt,runEvent:$t,transform:(e,t)=>t instanceof e.PyProxy?e.PyProxy.toJs(t):t,writeFile:({FS:e,_module:{PATH:t,PATH_FS:n}},r,s)=>gt({FS:e,PATH:t,PATH_FS:n},r,s)};const It={dict_converter:Object.fromEntries};var Mt={type:"pyodide",module:(e="0.24.1")=>`https://cdn.jsdelivr.net/pyodide/v${e}/full/pyodide.mjs`,async engine({loadPyodide:e},t,n){const{stderr:r,stdout:s,get:o}=dt(),a=n.slice(0,n.lastIndexOf("/")),i=await o(e({stderr:r,stdout:s,indexURL:a}));if(t.files&&await xt(this,i,t.files),t.fetch&&await vt(this,i,t.fetch),t.js_modules&&await At(t.js_modules),t.packages){await i.loadPackage("micropip");const e=await i.pyimport("micropip");await e.install(t.packages,{keep_going:!0}),e.destroy()}return i},registerJSModule:Ot,run:jt,runAsync:Rt,runEvent:$t,transform:(e,t)=>t instanceof e.ffi.PyProxy?t.toJs(It):t,writeFile:({FS:e,PATH:t,_module:{PATH_FS:n}},r,s)=>gt({FS:e,PATH:t,PATH_FS:n},r,s)};const Ft="ruby-wasm-wasi",Ht=Ft.replace(/\W+/g,"_");var Wt={type:Ft,experimental:!0,module:(e="2.2.0")=>`https://cdn.jsdelivr.net/npm/ruby-3_2-wasm-wasi@${e}/dist/browser.esm.js`,async engine({DefaultRubyVM:e},t,n){const r=await fetch(`${n.slice(0,n.lastIndexOf("/"))}/ruby.wasm`),s=await WebAssembly.compile(await r.arrayBuffer()),{vm:o}=await e(s);return t.files&&await xt(this,o,t.files),t.fetch&&await vt(this,o,t.fetch),t.js_modules&&await At(t.js_modules),o},registerJSModule(e,t,n){t=t.replace(/\W+/g,"__");const r=`__module_${Ht}_${t}`;globalThis[r]=n,this.run(e,`require "js";$${t}=JS.global[:${r}]`),delete globalThis[r]},run:(e,t,...n)=>e.eval(Je(t),...n),runAsync:(e,t,...n)=>e.evalAsync(Je(t),...n),async runEvent(e,t,n){if(/^xworker\.(on\w+)$/.test(t)){const{$1:t}=RegExp,r=`__module_${Ht}_event`;globalThis[r]=n,this.run(e,`require "js";$xworker.call("${t}",JS.global[:${r}])`),delete globalThis[r]}else{const r=this.run(e,`method(:${t})`);await r.call(t,e.wrap(n))}},transform:(e,t)=>t,writeFile:()=>{throw new Error(`writeFile is not supported in ${Ft}`)}};var Ct={type:"wasmoon",module:(e="1.16.0")=>`https://cdn.jsdelivr.net/npm/wasmoon@${e}/+esm`,async engine({LuaFactory:e,LuaLibraries:t},n){const{stderr:r,stdout:s,get:o}=dt(),a=await o((new e).createEngine());return a.global.getTable(t.Base,(e=>{a.global.setField(e,"print",s),a.global.setField(e,"printErr",r)})),n.files&&await xt(this,a,n.files),n.fetch&&await vt(this,a,n.fetch),n.js_modules&&await At(n.js_modules),a},registerJSModule:(e,t,n)=>{e.global.set(t,n)},run:(e,t,...n)=>{try{return e.doStringSync(Je(t),...n)}catch(t){ht.get(e).stderr(t)}},runAsync:async(e,t,...n)=>{try{return await e.doString(Je(t),...n)}catch(t){ht.get(e).stderr(t)}},runEvent:async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.global.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){ht.get(e).stderr(t)}},transform:(e,t)=>t,writeFile:({cmodule:{module:{FS:e}}},t,n)=>((e,t,n)=>(mt(e,yt(t)),t=wt(e,t),e.writeFile(t,new Uint8Array(n),{canOwn:!0})))(e,t,n)};const Lt=new Map,Dt=new Map,Jt=[],Bt=[],Ut=new Proxy(new Map,{get(e,t){if(!e.has(t)){const[n,...r]=t.split("@"),s=Lt.get(n),o=/^(?:\.?\.?\/|https?:\/\/)/i.test(r)?r.join("@"):s.module(...r);e.set(t,{url:o,module:import(o),engine:s.engine.bind(s)})}const{url:n,module:r,engine:s}=e.get(t);return(e,o)=>r.then((r=>{Dt.set(t,e);for(const t of["files","fetch"]){const n=e?.[t];n&&Et.set(n,o)}for(const t of["main","worker"]){const n=e?.js_modules?.[t];n&&Et.set(n,o)}return s(r,e,n)}))}}),qt=e=>{for(const t of[].concat(e.type))Lt.set(t,e),Jt.push(`script[type="${t}"]`),Bt.push(`${t}-`)};for(const e of[Nt,Mt,Wt,Ct])qt(e);const Gt=async e=>(await import("https://cdn.jsdelivr.net/npm/basic-toml@0.3.1/es.js")).parse(e),Kt=e=>{let t=typeof e;return"string"===t&&/\.(json|toml|txt)$/.test(e)?t=RegExp.$1:e="./config.txt",[Ze(e),t]},Xt=e=>{try{return JSON.parse(e)}catch(t){return Gt(e)}},Yt=(e,t,n={})=>{if(t){const[e,r]=Kt(t);"json"===r?n=fetch(e).then(Me):"toml"===r?n=fetch(e).then(Fe).then(Gt):"string"===r?n=Xt(t):"object"===r&&t?n=t:"txt"===r&&"string"==typeof n&&(n=Xt(n)),t=e}return Ve(n).then((n=>Ut[e](n,t)))},zt=(e,t="")=>`${e}@${t}`.replace(/@$/,""),Vt="BeforeRun",Zt="AfterRun",Qt=[`code${Vt}`,`code${Vt}Async`,`code${Zt}`,`code${Zt}Async`],en=["onWorker","onReady",`on${Vt}`,`on${Vt}Async`,`on${Zt}`,`on${Zt}Async`];function tn(e,t){const{run:n,runAsync:r}=Lt.get(this.type);return{...e,run:n.bind(this,t),runAsync:r.bind(this,t)}}const nn=(e,t,n,r,s,o)=>{if(s||o){const a=tn.bind(e,t),i=r?"runAsync":"run",c=e[i];e[i]=r?async function(e,t,...r){s&&await s.call(this,a(e),n);const i=await c.call(this,e,t,...r);return o&&await o.call(this,a(e),n),i}:function(e,t,...r){s&&s.call(this,a(e),n);const i=c.call(this,e,t,...r);return o&&o.call(this,a(e),n),i}}};let rn=class{constructor(e,t={}){const{main:n,worker:r}=t;this.interpreter=e,this.onWorker=n?.onWorker;for(const e of en.slice(1))this[e]=r?.[e];for(const e of Qt)this[e]=r?.[e]}toJSON(){const e={};for(const t of en.slice(1))this[t]&&(e[t]=String(this[t]));for(const t of Qt)this[t]&&(e[t]=Je(this[t]()));return e}};var sn=(...e)=>function(t,n){const r=new Worker(URL.createObjectURL(new Blob(['const e="object"==typeof self?self:globalThis,t=t=>((t,n)=>{const r=(e,n)=>(t.set(n,e),e),s=o=>{if(t.has(o))return t.get(o);const[a,i]=n[o];switch(a){case 0:case-1:return r(i,o);case 1:{const e=r([],o);for(const t of i)e.push(s(t));return e}case 2:{const e=r({},o);for(const[t,n]of i)e[s(t)]=s(n);return e}case 3:return r(new Date(i),o);case 4:{const{source:e,flags:t}=i;return r(new RegExp(e,t),o)}case 5:{const e=r(new Map,o);for(const[t,n]of i)e.set(s(t),s(n));return e}case 6:{const e=r(new Set,o);for(const t of i)e.add(s(t));return e}case 7:{const{name:t,message:n}=i;return r(new e[t](n),o)}case 8:return r(BigInt(i),o);case"BigInt":return r(Object(BigInt(i)),o)}return r(new e[a](i),o)};return s})(new Map,t)(0),n="",{toString:r}={},{keys:s}=Object,o=e=>{const t=typeof e;if("object"!==t||!e)return[0,t];const s=r.call(e).slice(8,-1);switch(s){case"Array":return[1,n];case"Object":return[2,n];case"Date":return[3,n];case"RegExp":return[4,n];case"Map":return[5,n];case"Set":return[6,n]}return s.includes("Array")?[1,s]:s.includes("Error")?[7,s]:[2,s]},a=([e,t])=>0===e&&("function"===t||"symbol"===t),i=(e,{json:t,lossy:n}={})=>{const r=[];return((e,t,n,r)=>{const i=(e,t)=>{const s=r.push(e)-1;return n.set(t,s),s},c=r=>{if(n.has(r))return n.get(r);let[l,u]=o(r);switch(l){case 0:{let t=r;switch(u){case"bigint":l=8,t=r.toString();break;case"function":case"symbol":if(e)throw new TypeError("unable to serialize "+u);t=null;break;case"undefined":return i([-1],r)}return i([l,t],r)}case 1:{if(u)return i([u,[...r]],r);const e=[],t=i([l,e],r);for(const t of r)e.push(c(t));return t}case 2:{if(u)switch(u){case"BigInt":return i([u,r.toString()],r);case"Boolean":case"Number":case"String":return i([u,r.valueOf()],r)}if(t&&"toJSON"in r)return c(r.toJSON());const n=[],p=i([l,n],r);for(const t of s(r))!e&&a(o(r[t]))||n.push([c(t),c(r[t])]);return p}case 3:return i([l,r.toISOString()],r);case 4:{const{source:e,flags:t}=r;return i([l,{source:e,flags:t}],r)}case 5:{const t=[],n=i([l,t],r);for(const[n,s]of r)(e||!a(o(n))&&!a(o(s)))&&t.push([c(n),c(s)]);return n}case 6:{const t=[],n=i([l,t],r);for(const n of r)!e&&a(o(n))||t.push(c(n));return n}}const{message:p}=r;return i([l,{name:u,message:p}],r)};return c})(!(t||n),!!t,new Map,r)(e),r},{parse:c,stringify:l}=JSON,u={json:!0,lossy:!0};var p=Object.freeze({__proto__:null,parse:e=>t(c(e)),stringify:e=>l(i(e,u))});const f="ac6c1995-0675-46af-a99f-7fa298fa4e6d",h="M"+f,d="T"+f,g="array",y="function",m="null",w="number",_="object",b="string",E="symbol",k="undefined",v="apply",T="construct",S="defineProperty",x="deleteProperty",A="get",O="getOwnPropertyDescriptor",j="getPrototypeOf",R="has",P="isExtensible",N="ownKeys",I="preventExtensions",$="set",M="setPrototypeOf",F="delete";var H=e=>({value:new Promise((t=>{let n=new Worker("data:application/javascript,onmessage%3D(%7Bdata%3Ab%7D)%3D%3E(Atomics.wait(b%2C0)%2CpostMessage(0))");n.onmessage=t,n.postMessage(e)}))})\n/*! (c) Andrea Giammarchi - ISC */;const{Int32Array:W,Map:C,SharedArrayBuffer:D,Uint16Array:L}=globalThis,{BYTES_PER_ELEMENT:B}=W,{BYTES_PER_ELEMENT:U}=L,{isArray:J}=Array,{notify:q,wait:G,waitAsync:K}=Atomics,Y=new WeakSet,z=new WeakMap,X={value:{then:e=>e()}};let V=0;const Z=(e,{parse:t=JSON.parse,stringify:n=JSON.stringify,transform:r,interrupt:s}=JSON)=>{if(!z.has(e)){const o=(t,...n)=>e.postMessage({[f]:n},{transfer:t}),a=typeof s===y?s:s?.handler,i=s?.delay||42,c=new TextDecoder("utf-16"),l=(e,t)=>e?(K||H)(t,0):(a?((e,t,n)=>{for(;"timed-out"===G(e,0,0,t);)n()})(t,i,a):G(t,0),X);let u=!1;z.set(e,new Proxy(new C,{[R]:(e,t)=>"string"==typeof t&&!t.startsWith("_"),[A]:(n,s)=>"then"===s?null:(...n)=>{const a=V++;let i=new W(new D(2*B)),p=[];Y.has(n.at(-1)||p)&&Y.delete(p=n.pop()),o(p,a,i,s,r?n.map(r):n);const f=e!==globalThis;let h=0;return u&&f&&(h=setTimeout(console.warn,1e3,`💀🔒 - Possible deadlock if proxy.${s}(...args) is awaited`)),l(f,i).value.then((()=>{clearTimeout(h);const e=i[1];if(!e)return;const n=U*e;return i=new W(new D(n+n%B)),o([],a,i),l(f,i).value.then((()=>t(c.decode(new L(i.buffer).slice(0,e)))))}))},[$](t,s,o){const a=typeof o;if(a!==y)throw new Error(`Unable to assign ${s} as ${a}`);if(!t.size){const s=new C;e.addEventListener("message",(async e=>{const o=e.data?.[f];if(J(o)){e.stopImmediatePropagation();const[a,i,...c]=o;let l;if(c.length){const[e,o]=c;if(t.has(e)){u=!0;try{const c=await t.get(e)(...o);if(void 0!==c){const e=n(r?r(c):c);s.set(a,e),i[1]=e.length}}catch(e){l=e}finally{u=!1}}else l=new Error(`Unsupported action: ${e}`);i[0]=1}else{const e=s.get(a);s.delete(a);for(let t=new L(i.buffer),n=0;n<e.length;n++)t[n]=e.charCodeAt(n)}if(q(i,0),l)throw l}}))}return!!t.set(s,o)}}))}return z.get(e)};Z.transfer=(...e)=>(Y.add(e),e);const{isArray:Q}=Array,ee=(e,t)=>t,te=e=>typeof e===y?(e=>e())(e):e;function ne(){return this}const re=(e,t)=>e===g?[t]:{t:e,v:t},se=(e,t=ee)=>{let n=typeof e,r=e;return n===_&&(Q(e)?(n=g,r=e.at(0)):({t:n,v:r}=e)),t(n,r)},oe=(e,t)=>e===y?t:re(e,t),ae=(e,t=oe)=>{const n=null===e?m:typeof e;return t(n===_&&Q(e)?g:n,e)},ie=new FinalizationRegistry((([e,t,n])=>{n&&console.debug(`Held value ${String(t)} not relevant anymore`),e(t)})),ce=Object.create(null),le=(e,t,{debug:n,return:r,token:s=e}=ce)=>{const o=r||new Proxy(e,ce),a=[o,[t,e,!!n]];return!1!==s&&a.push(s),ie.register(...a),o},{defineProperty:ue,deleteProperty:pe,getOwnPropertyDescriptor:fe,getPrototypeOf:he,isExtensible:de,ownKeys:ge,preventExtensions:ye,set:me,setPrototypeOf:we}=Reflect,{assign:_e,create:be}=Object,Ee=he(Int8Array),ke=(e,t)=>{const{get:n,set:r,value:s}=e;return n&&(e.get=t(n)),r&&(e.set=t(r)),s&&(e.value=t(s)),e},ve=e=>t=>ae(t,((t,n)=>{switch(t){case m:return re(m,n);case _:if(n===globalThis)return re(t,null);case g:case y:return e(t,n);case"boolean":case w:case b:case k:case"bigint":return re(t,n);case E:{if(Te.has(n))return re(t,Te.get(n));let e=Symbol.keyFor(n);if(e)return re(t,`.${e}`)}}throw new TypeError(`Unable to handle this ${t}: ${String(n)}`)})),Te=new Map(ge(Symbol).filter((e=>typeof Symbol[e]===E)).map((e=>[Symbol[e],e]))),Se=e=>{if(e.startsWith("."))return Symbol.for(e.slice(1));for(const[t,n]of Te)if(n===e)return t},xe=e=>e;var Ae=((e,t)=>{const n=t&&new WeakMap;if(t){const{addEventListener:e}=EventTarget.prototype;ue(EventTarget.prototype,"addEventListener",{value(t,r,...s){return s.at(0)?.invoke&&(n.has(this)||n.set(this,new Map),n.get(this).set(t,[].concat(s[0].invoke)),delete s[0].invoke),e.call(this,t,r,...s)}})}const r=t&&(e=>{const{currentTarget:t,target:r,type:s}=e;for(const o of n.get(t||r)?.get(s)||[])e[o]()});return function(n,s,o,...a){let i=0,c=this?.transform||xe;const l=new Map,u=new Map,{[o]:p}=n,f=a.length?_e(be(globalThis),...a):globalThis,h=ve(((e,t)=>{if(!l.has(t)){let n;for(;u.has(n=i++););l.set(t,n),u.set(n,e===y?t:c(t))}return re(e,l.get(t))})),d=e=>{p(F,re(b,e))},m=(e,n)=>{switch(e){case _:if(null==n)return f;case g:if(typeof n===w)return u.get(n);if(!(n instanceof Ee))for(const e in n)n[e]=H(n[e]);return n;case y:if(typeof n===b){if(!u.has(n)){const e=function(...e){return t&&e.at(0)instanceof Event&&r(...e),p(v,re(y,n),h(this),e.map(h))};return u.set(n,new WeakRef(e)),le(n,d,{return:e,token:!1})}return u.get(n).deref()}return u.get(n);case E:return Se(n)}return n},H=e=>se(e,m),W={[v]:(e,t,n)=>h(e.apply(t,n)),[T]:(e,t)=>h(new e(...t)),[S]:(e,t,n)=>h(ue(e,t,n)),[x]:(e,t)=>h(pe(e,t)),[j]:e=>h(he(e)),[A]:(e,t)=>h(e[t]),[O]:(e,t)=>{const n=fe(e,t);return n?re(_,ke(n,h)):re(k,n)},[R]:(e,t)=>h(t in e),[P]:e=>h(de(e)),[N]:e=>re(g,ge(e).map(h)),[I]:e=>h(ye(e)),[$]:(e,t,n)=>h(me(e,t,n)),[M]:(e,t)=>h(we(e,t)),[F](e){l.delete(u.get(e)),u.delete(e)}};return n[s]=(e,t,...n)=>{switch(e){case v:n[0]=H(n[0]),n[1]=n[1].map(H);break;case T:n[0]=n[0].map(H);break;case S:{const[e,t]=n;n[0]=H(e);const{get:r,set:s,value:o}=t;r&&(t.get=H(r)),s&&(t.set=H(s)),o&&(t.value=H(o));break}default:n=n.map(H)}return W[e](H(t),...n)},{proxy:n,[e.toLowerCase()]:f,[`is${e}Proxy`]:()=>!1}}})("Window",!0),Oe=(e=>{let t=0;const n=new Map,r=new Map,s=Symbol();return function(o,a,i){const c=this?.transform||xe,{[a]:l}=o,u=new Map,p=e=>{u.delete(e),l(F,f(e))},f=ve(((e,o)=>{if(s in o)return te(o[s]);if(e===y){if(o=c(o),!r.has(o)){let e;for(;r.has(e=String(t++)););n.set(o,e),r.set(e,o)}return re(e,n.get(o))}if(!(o instanceof Ee)){o=c(o);for(const e in o)o[e]=f(o[e])}return re(e,o)})),h=(e,t,n)=>{if(!u.has(n)){const r=t===y?(e=>ne.bind(e))(e):e,s=new Proxy(r,k);return u.set(n,new WeakRef(s)),le(n,p,{return:s,token:!1})}return u.get(n).deref()},d=e=>se(e,((t,n)=>{switch(t){case _:if(null===n)return globalThis;case g:return typeof n===w?h(e,t,n):n;case y:return typeof n===b?r.get(n):h(e,t,n);case E:return Se(n)}return n})),m=(e,t,...n)=>d(l(e,te(t),...n)),k={[v]:(e,t,n)=>m(v,e,f(t),n.map(f)),[T]:(e,t)=>m(T,e,t.map(f)),[S]:(e,t,n)=>{const{get:r,set:s,value:o}=n;return typeof r===y&&(n.get=f(r)),typeof s===y&&(n.set=f(s)),typeof o===y&&(n.value=f(o)),m(S,e,f(t),n)},[x]:(e,t)=>m(x,e,f(t)),[j]:e=>m(j,e),[A]:(e,t)=>t===s?e:m(A,e,f(t)),[O]:(e,t)=>{const n=m(O,e,f(t));return n&&ke(n,d)},[R]:(e,t)=>t===s||m(R,e,f(t)),[P]:e=>m(P,e),[N]:e=>m(N,e).map(d),[I]:e=>m(I,e),[$]:(e,t,n)=>m($,e,f(t),f(n)),[M]:(e,t)=>m(M,e,f(t))};o[i]=(e,t,s,o)=>{switch(e){case v:return d(t).apply(d(s),o.map(d));case F:{const e=d(t);n.delete(r.get(e)),r.delete(e)}}};const H=new Proxy(re(_,null),k);return{[e.toLowerCase()]:H,[`is${e}Proxy`]:e=>typeof e===_&&!!e&&s in e,proxy:o}}})("Window"),je=typeof Worker===y?Worker:class{};const Re=new WeakMap,Pe=(e,...t)=>{const n=Z(e,...t);if(!Re.has(n)){const r=e instanceof je?Ae:Oe;Re.set(n,r.call(t.at(0),n,h,d))}return Re.get(n)};Pe.transfer=Z.transfer;const Ne={object(...e){return this.string(function(e){for(var t=e[0],n=1,r=arguments.length;n<r;n++)t+=arguments[n]+e[n];return t}(...e))},string(e){for(const t of e.split(/[\\r\\n]+/))if(t.trim().length){/^(\\s+)/.test(t)&&(e=e.replace(new RegExp("^"+RegExp.$1,"gm"),""));break}return e}};Promise.withResolvers||(Promise.withResolvers=function(){var e,t,n=new this((function(n,r){e=n,t=r}));return{resolve:e,reject:t,promise:n}});const Ie=e=>e.arrayBuffer(),$e=e=>e.json(),Me=e=>e.text(),Fe=new WeakMap,He=e=>{const t=e||console,n={stderr:(t.stderr||console.error).bind(t),stdout:(t.stdout||console.log).bind(t)};return{stderr:(...e)=>n.stderr(...e),stdout:(...e)=>n.stdout(...e),async get(e){const t=await e;return Fe.set(t,n),t}}},We=({FS:e,PATH:t,PATH_FS:n},r,s)=>{const o=n.resolve(r);return e.mkdirTree(t.dirname(o)),e.writeFile(o,new Uint8Array(s),{canOwn:!0})},Ce=e=>{const t=e.split("/");return t.pop(),t.join("/")},De=(e,t)=>{const n=[];for(const r of t.split("/"))"."!==r&&(n.push(r),r&&e.mkdir(n.join("/")))},Le=(e,t)=>{const n=[];for(const e of t.split("/"))switch(e){case"":case".":break;case"..":n.pop();break;default:n.push(e)}return[e.cwd()].concat(n).join("/").replace(/^\\/+/,"/")},Be=e=>{const t=e.map((e=>e.trim().replace(/(^[/]*|[/]*$)/g,""))).filter((e=>""!==e&&"."!==e)).join("/");return e[0].startsWith("/")?`/${t}`:t},Ue=(e,t)=>fetch(at(t,Je.get(e))),Je=new WeakMap,qe=(e,t,n)=>st((e=>{for(const{files:t,to_file:n,from:r=""}of e){if(void 0!==t&&void 0!==n)throw new Error("Cannot use \'to_file\' and \'files\' parameters together!");if(void 0===t&&void 0===n&&r.endsWith("/"))throw new Error(`Couldn\'t determine the filename from the path ${r}, please supply \'to_file\' parameter.`)}return e.flatMap((({from:e="",to_folder:t=".",to_file:n,files:r})=>{if(Ze(r))return r.map((n=>({url:Be([e,n]),path:Be([t,n])})));const s=n||e.slice(1+e.lastIndexOf("/"));return[{url:e,path:Be([t,s])}]}))})(n).map((({url:r,path:s})=>Ue(n,r).then(Ie).then((n=>e.writeFile(t,s,n)))))),Ge=(e,t)=>t.endsWith("/")?`${t}${e.split("/").pop()}`:t,Ke=(e,t)=>e.replace(/\\{.+?\\}/g,(e=>{if(!t.has(e))throw new SyntaxError(`Invalid template: ${e}`);return t.get(e)})),Ye=(e,t,n)=>st((e=>{const t=new Map,n=new Set,r=[];for(const[s,o]of rt(e))if(/^\\{.+\\}$/.test(s)){if(t.has(s))throw new SyntaxError(`Duplicated template: ${s}`);t.set(s,Ke(o,t))}else{const e=Ke(s,t),a=Ge(e,Ke(o||"./",t));if(n.has(a))throw new SyntaxError(`Duplicated destination: ${a}`);n.add(a),r.push({url:e,path:a})}return r})(n).map((({url:r,path:s})=>Ue(n,r).then(Ie).then((n=>e.writeFile(t,s,n)))))),ze=typeof document===k,Xe=({main:e,worker:t})=>{const n=[];if(t&&ze)for(let[e,r]of rt(t))e=at(e,Je.get(t)),n.push(ht(e,r));if(e&&!ze)for(let[t,r]of rt(e))t=at(t,Je.get(e)),gt(t)?dt(t):n.push(ht(t,r));return st(n)},Ve=(e,...t)=>Ne[typeof e](e,...t),{isArray:Ze}=Array,{assign:Qe,create:et,defineProperties:tt,defineProperty:nt,entries:rt}=Object,{all:st,resolve:ot}=new Proxy(Promise,{get:(e,t)=>e[t].bind(e)}),at=(e,t=location.href)=>new URL(e,t.replace(/^blob:/,"")).href,it=(e,t,n,r=!1,s=CustomEvent)=>{e.dispatchEvent(new s(`${t}:${n}`,{bubbles:!0,detail:{worker:r}}))},ct=e=>Function(`\'use strict\';return (${e})`)(),lt=Symbol.for("polyscript.js_modules"),ut=new Map;nt(globalThis,lt,{value:ut}),new Proxy(ut,{get:(e,t)=>e.get(t),has:(e,t)=>e.has(t),ownKeys:e=>[...e.keys()]});const pt=(e,t)=>!t.startsWith("_"),ft=(e,t)=>new Proxy(e,{has:pt,get:(e,n)=>e[t][n]}),ht=(e,t)=>import(e).then((e=>{ut.set(t,{...e})})),dt=e=>new Promise(((t,n)=>{document.querySelector(`link[href="${e}"]`)&&t(),document.head.append(Qe(document.createElement("link"),{rel:"stylesheet",href:e,onload:t,onerror:n}))})),gt=e=>/\\.css/i.test(new URL(e).pathname),yt=(e,t)=>e.has(t),mt=e=>[...e.keys()];var wt=(e,t,n)=>{let r=globalThis[lt],s="";if(n){s=Je.get(n);for(let[e,t]of rt(n)){let n=r.get(t);n&&!Ze(n)||(r.set(t,n||(n=[])),n.push(e))}}return((e,t,n,r)=>new Proxy(e,{has:yt,ownKeys:mt,get:(e,s)=>{let o=e.get(s);if(Ze(o)){let a=o;o=null;for(let e of a)e=at(e,r),gt(e)?n.importCSS(e):(n.importJS(e,s),o=t[lt].get(s));e.set(s,o)}return o}}))(r,e,t,s)};const _t=(e,t,n)=>{e.registerJsModule(t,n)},bt=(e,t,...n)=>{try{return e.runPython(Ve(t),...n)}catch(t){Fe.get(e).stderr(t)}},Et=async(e,t,...n)=>{try{return await e.runPythonAsync(Ve(t),...n)}catch(t){Fe.get(e).stderr(t)}},kt=async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.globals.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){Fe.get(e).stderr(t)}};var vt=(new TextEncoder).encode(\'from uio import StringIO\\nimport sys\\n\\nclass Response:\\n    def __init__(self, f):\\n        self.raw = f\\n        self.encoding = "utf-8"\\n        self._cached = None\\n\\n    def close(self):\\n        if self.raw:\\n            self.raw.close()\\n            self.raw = None\\n        self._cached = None\\n\\n    @property\\n    def content(self):\\n        if self._cached is None:\\n            try:\\n                self._cached = self.raw.read()\\n            finally:\\n                self.raw.close()\\n                self.raw = None\\n        return self._cached\\n\\n    @property\\n    def text(self):\\n        return str(self.content, self.encoding)\\n\\n    def json(self):\\n        import ujson\\n\\n        return ujson.loads(self.content)\\n\\n\\n# TODO try to support streaming xhr requests, a-la pyodide-http\\nHEADERS_TO_IGNORE = ("user-agent",)\\n\\n\\ntry:\\n    import js\\nexcept Exception as err:\\n    raise OSError("This version of urequests can only be used in the browser")\\n\\n# TODO try to support streaming xhr requests, a-la pyodide-http\\n\\nHEADERS_TO_IGNORE = ("user-agent",)\\n\\n\\ndef request(\\n    method,\\n    url,\\n    data=None,\\n    json=None,\\n    headers={},\\n    stream=None,\\n    auth=None,\\n    timeout=None,\\n    parse_headers=True,\\n):\\n    from js import XMLHttpRequest\\n\\n    xhr = XMLHttpRequest.new()\\n    xhr.withCredentials = False\\n\\n    if auth is not None:\\n        import ubinascii\\n\\n        username, password = auth\\n        xhr.open(method, url, False, username, password)\\n    else:\\n        xhr.open(method, url, False)\\n\\n    for name, value in headers.items():\\n        if name.lower() not in HEADERS_TO_IGNORE:\\n            xhr.setRequestHeader(name, value)\\n\\n    if timeout:\\n        xhr.timeout = int(timeout * 1000)\\n\\n    if json is not None:\\n        assert data is None\\n        import ujson\\n\\n        data = ujson.dumps(json)\\n        # s.write(b"Content-Type: application/json\\\\r\\\\n")\\n        xhr.setRequestHeader("Content-Type", "application/json")\\n\\n    xhr.send(data)\\n\\n    # Emulates the construction process in the original urequests\\n    resp = Response(StringIO(xhr.responseText))\\n    resp.status_code = xhr.status\\n    resp.reason = xhr.statusText\\n    resp.headers = xhr.getAllResponseHeaders()\\n\\n    return resp\\n\\n\\n# Other methods - head, post, put, patch, delete - are not used by\\n# mip and therefore not included\\n\\n\\ndef get(url, **kw):\\n    return request("GET", url, **kw)\\n\\n\\n# Content below this line is from the Micropython MIP package and is covered\\n# by the applicable MIT license:\\n# \\n# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\\n# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \\n# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\\n# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \\n# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \\n# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER \\n# DEALINGS IN THE SOFTWARE.\\n\\n# MicroPython package installer\\n# MIT license; Copyright (c) 2022 Jim Mussared\\n\\n\\n_PACKAGE_INDEX = const("https://micropython.org/pi/v2")\\n_CHUNK_SIZE = 128\\n\\n\\n# This implements os.makedirs(os.dirname(path))\\ndef _ensure_path_exists(path):\\n    import os\\n\\n    split = path.split("/")\\n\\n    # Handle paths starting with "/".\\n    if not split[0]:\\n        split.pop(0)\\n        split[0] = "/" + split[0]\\n\\n    prefix = ""\\n    for i in range(len(split) - 1):\\n        prefix += split[i]\\n        try:\\n            os.stat(prefix)\\n        except:\\n            os.mkdir(prefix)\\n        prefix += "/"\\n\\n\\n# Copy from src (stream) to dest (function-taking-bytes)\\ndef _chunk(src, dest):\\n    buf = memoryview(bytearray(_CHUNK_SIZE))\\n    while True:\\n        n = src.readinto(buf)\\n        if n == 0:\\n            break\\n        dest(buf if n == _CHUNK_SIZE else buf[:n])\\n\\n\\n# Check if the specified path exists and matches the hash.\\ndef _check_exists(path, short_hash):\\n    import os\\n\\n    try:\\n        import binascii\\n        import hashlib\\n\\n        with open(path, "rb") as f:\\n            hs256 = hashlib.sha256()\\n            _chunk(f, hs256.update)\\n            existing_hash = str(binascii.hexlify(hs256.digest())[: len(short_hash)], "utf-8")\\n            return existing_hash == short_hash\\n    except:\\n        return False\\n\\n\\ndef _rewrite_url(url, branch=None):\\n    if not branch:\\n        branch = "HEAD"\\n    if url.startswith("github:"):\\n        url = url[7:].split("/")\\n        url = (\\n            "https://raw.githubusercontent.com/"\\n            + url[0]\\n            + "/"\\n            + url[1]\\n            + "/"\\n            + branch\\n            + "/"\\n            + "/".join(url[2:])\\n        )\\n    return url\\n\\n\\ndef _download_file(url, dest):\\n    response = get(url)\\n    try:\\n        if response.status_code != 200:\\n            print("Error", response.status_code, "requesting", url)\\n            return False\\n\\n        print("Copying:", dest)\\n        _ensure_path_exists(dest)\\n        with open(dest, "wb") as f:\\n            _chunk(response.raw, f.write)\\n\\n        return True\\n    finally:\\n        response.close()\\n\\n\\ndef _install_json(package_json_url, index, target, version, mpy):\\n    response = get(_rewrite_url(package_json_url, version))\\n    try:\\n        if response.status_code != 200:\\n            print("Package not found:", package_json_url)\\n            return False\\n\\n        package_json = response.json()\\n    finally:\\n        response.close()\\n    for target_path, short_hash in package_json.get("hashes", ()):\\n        fs_target_path = target + "/" + target_path\\n        if _check_exists(fs_target_path, short_hash):\\n            print("Exists:", fs_target_path)\\n        else:\\n            file_url = "{}/file/{}/{}".format(index, short_hash[:2], short_hash)\\n            if not _download_file(file_url, fs_target_path):\\n                print("File not found: {} {}".format(target_path, short_hash))\\n                return False\\n    for target_path, url in package_json.get("urls", ()):\\n        fs_target_path = target + "/" + target_path\\n        if not _download_file(_rewrite_url(url, version), fs_target_path):\\n            print("File not found: {} {}".format(target_path, url))\\n            return False\\n    for dep, dep_version in package_json.get("deps", ()):\\n        if not _install_package(dep, index, target, dep_version, mpy):\\n            return False\\n    return True\\n\\n\\ndef _install_package(package, index, target, version, mpy):\\n    if (\\n        package.startswith("http://")\\n        or package.startswith("https://")\\n        or package.startswith("github:")\\n    ):\\n        if package.endswith(".py") or package.endswith(".mpy"):\\n            print("Downloading {} to {}".format(package, target))\\n            return _download_file(\\n                _rewrite_url(package, version), target + "/" + package.rsplit("/")[-1]\\n            )\\n        else:\\n            if not package.endswith(".json"):\\n                if not package.endswith("/"):\\n                    package += "/"\\n                package += "package.json"\\n            print("Installing {} to {}".format(package, target))\\n    else:\\n        if not version:\\n            version = "latest"\\n        print("Installing {} ({}) from {} to {}".format(package, version, index, target))\\n\\n        mpy_version = (\\n            sys.implementation._mpy & 0xFF if mpy and hasattr(sys.implementation, "_mpy") else "py"\\n        )\\n\\n        package = "{}/package/{}/{}/{}.json".format(index, mpy_version, package, version)\\n\\n    return _install_json(package, index, target, version, mpy)\\n\\n\\ndef install(package, index=None, target=None, version=None, mpy=True):\\n    if not target:\\n        for p in sys.path:\\n            if p.endswith("/lib"):\\n                target = p\\n                break\\n        else:\\n            print("Unable to find lib dir in sys.path")\\n            return\\n\\n    if not index:\\n        index = _PACKAGE_INDEX\\n\\n    if _install_package(package, index.rstrip("/"), target, version, mpy):\\n        print("Done")\\n    else:\\n        print("Package may be partially installed")\\n\');var Tt={type:"micropython",module:(e="1.21.0-279")=>`https://cdn.jsdelivr.net/npm/@micropython/micropython-webassembly-pyscript@${e}/micropython.mjs`,async engine({loadMicroPython:e},t,n){const{stderr:r,stdout:s,get:o}=He();n=n.replace(/\\.m?js$/,".wasm");const a=await o(e({stderr:r,stdout:s,url:n}));if(t.files&&await Ye(this,a,t.files),t.fetch&&await qe(this,a,t.fetch),t.js_modules&&await Xe(t.js_modules),this.writeFile(a,"./mip.py",vt),t.packages){const e=a.pyimport("mip");for(const n of t.packages)e.install(n)}return a},registerJSModule:_t,run:bt,runAsync:Et,runEvent:kt,transform:(e,t)=>t instanceof e.PyProxy?e.PyProxy.toJs(t):t,writeFile:({FS:e,_module:{PATH:t,PATH_FS:n}},r,s)=>We({FS:e,PATH:t,PATH_FS:n},r,s)};const St={dict_converter:Object.fromEntries};var xt={type:"pyodide",module:(e="0.24.1")=>`https://cdn.jsdelivr.net/pyodide/v${e}/full/pyodide.mjs`,async engine({loadPyodide:e},t,n){const{stderr:r,stdout:s,get:o}=He(),a=n.slice(0,n.lastIndexOf("/")),i=await o(e({stderr:r,stdout:s,indexURL:a}));if(t.files&&await Ye(this,i,t.files),t.fetch&&await qe(this,i,t.fetch),t.js_modules&&await Xe(t.js_modules),t.packages){await i.loadPackage("micropip");const e=await i.pyimport("micropip");await e.install(t.packages,{keep_going:!0}),e.destroy()}return i},registerJSModule:_t,run:bt,runAsync:Et,runEvent:kt,transform:(e,t)=>t instanceof e.ffi.PyProxy?t.toJs(St):t,writeFile:({FS:e,PATH:t,_module:{PATH_FS:n}},r,s)=>We({FS:e,PATH:t,PATH_FS:n},r,s)};const At="ruby-wasm-wasi",Ot=At.replace(/\\W+/g,"_");var jt={type:At,experimental:!0,module:(e="2.2.0")=>`https://cdn.jsdelivr.net/npm/ruby-3_2-wasm-wasi@${e}/dist/browser.esm.js`,async engine({DefaultRubyVM:e},t,n){const r=await fetch(`${n.slice(0,n.lastIndexOf("/"))}/ruby.wasm`),s=await WebAssembly.compile(await r.arrayBuffer()),{vm:o}=await e(s);return t.files&&await Ye(this,o,t.files),t.fetch&&await qe(this,o,t.fetch),t.js_modules&&await Xe(t.js_modules),o},registerJSModule(e,t,n){t=t.replace(/\\W+/g,"__");const r=`__module_${Ot}_${t}`;globalThis[r]=n,this.run(e,`require "js";$${t}=JS.global[:${r}]`),delete globalThis[r]},run:(e,t,...n)=>e.eval(Ve(t),...n),runAsync:(e,t,...n)=>e.evalAsync(Ve(t),...n),async runEvent(e,t,n){if(/^xworker\\.(on\\w+)$/.test(t)){const{$1:t}=RegExp,r=`__module_${Ot}_event`;globalThis[r]=n,this.run(e,`require "js";$xworker.call("${t}",JS.global[:${r}])`),delete globalThis[r]}else{const r=this.run(e,`method(:${t})`);await r.call(t,e.wrap(n))}},transform:(e,t)=>t,writeFile:()=>{throw new Error(`writeFile is not supported in ${At}`)}};var Rt={type:"wasmoon",module:(e="1.16.0")=>`https://cdn.jsdelivr.net/npm/wasmoon@${e}/+esm`,async engine({LuaFactory:e,LuaLibraries:t},n){const{stderr:r,stdout:s,get:o}=He(),a=await o((new e).createEngine());return a.global.getTable(t.Base,(e=>{a.global.setField(e,"print",s),a.global.setField(e,"printErr",r)})),n.files&&await Ye(this,a,n.files),n.fetch&&await qe(this,a,n.fetch),n.js_modules&&await Xe(n.js_modules),a},registerJSModule:(e,t,n)=>{e.global.set(t,n)},run:(e,t,...n)=>{try{return e.doStringSync(Ve(t),...n)}catch(t){Fe.get(e).stderr(t)}},runAsync:async(e,t,...n)=>{try{return await e.doString(Ve(t),...n)}catch(t){Fe.get(e).stderr(t)}},runEvent:async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.global.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){Fe.get(e).stderr(t)}},transform:(e,t)=>t,writeFile:({cmodule:{module:{FS:e}}},t,n)=>((e,t,n)=>(De(e,Ce(t)),t=Le(e,t),e.writeFile(t,new Uint8Array(n),{canOwn:!0})))(e,t,n)};const Pt=new Map,Nt=new Map,It=new Proxy(new Map,{get(e,t){if(!e.has(t)){const[n,...r]=t.split("@"),s=Pt.get(n),o=/^(?:\\.?\\.?\\/|https?:\\/\\/)/i.test(r)?r.join("@"):s.module(...r);e.set(t,{url:o,module:import(o),engine:s.engine.bind(s)})}const{url:n,module:r,engine:s}=e.get(t);return(e,o)=>r.then((r=>{Nt.set(t,e);for(const t of["files","fetch"]){const n=e?.[t];n&&Je.set(n,o)}for(const t of["main","worker"]){const n=e?.js_modules?.[t];n&&Je.set(n,o)}return s(r,e,n)}))}}),$t=e=>{for(const t of[].concat(e.type))Pt.set(t,e)};for(const e of[Tt,xt,jt,Rt])$t(e);const Mt=async e=>(await import("https://cdn.jsdelivr.net/npm/basic-toml@0.3.1/es.js")).parse(e),Ft=e=>{try{return JSON.parse(e)}catch(t){return Mt(e)}},Ht=(e,t,n={})=>{if(t){const[e,r]=(e=>{let t=typeof e;return"string"===t&&/\\.(json|toml|txt)$/.test(e)?t=RegExp.$1:e="./config.txt",[at(e),t]})(t);"json"===r?n=fetch(e).then($e):"toml"===r?n=fetch(e).then(Me).then(Mt):"string"===r?n=Ft(t):"object"===r&&t?n=t:"txt"===r&&"string"==typeof n&&(n=Ft(n)),t=e}return ot(n).then((n=>It[e](n,t)))},Wt="BeforeRun",Ct="AfterRun",Dt=[`code${Wt}`,`code${Wt}Async`,`code${Ct}`,`code${Ct}Async`],Lt=["onWorker","onReady",`on${Wt}`,`on${Wt}Async`,`on${Ct}`,`on${Ct}Async`];function Bt(e,t){const{run:n,runAsync:r}=Pt.get(this.type);return{...e,run:n.bind(this,t),runAsync:r.bind(this,t)}}const Ut=(e,t,n,r,s,o)=>{if(s||o){const a=Bt.bind(e,t),i=r?"runAsync":"run",c=e[i];e[i]=r?async function(e,t,...r){s&&await s.call(this,a(e),n);const i=await c.call(this,e,t,...r);return o&&await o.call(this,a(e),n),i}:function(e,t,...r){s&&s.call(this,a(e),n);const i=c.call(this,e,t,...r);return o&&o.call(this,a(e),n),i}}};try{new SharedArrayBuffer(4)}catch(e){throw new Error(["Unable to use SharedArrayBuffer due insecure environment.","Please read requirements in MDN: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer#security_requirements"].join("\\n"))}let Jt,qt,Gt;const Kt=(e,t)=>{addEventListener(e,t||(async t=>{try{await Jt,qt(`xworker.on${e}`,t)}catch(e){postMessage(e)}}),!!t&&{once:!0})},{parse:Yt,stringify:zt}=p,{proxy:Xt,window:Vt,isWindowProxy:Zt}=Pe(self,{parse:Yt,stringify:zt,transform:e=>Gt?Gt(e):e}),Qt={sync:Xt,window:Vt,isWindowProxy:Zt,onmessage:console.info,onerror:console.error,onmessageerror:console.warn,postMessage:postMessage.bind(self)};Kt("message",(({data:{options:e,config:t,code:n,hooks:r}})=>{Jt=(async()=>{try{const{id:s,tag:o,type:a,custom:i,version:c,config:l,async:u}=e,p=((e,t="")=>`${e}@${t}`.replace(/@$/,""))(a,c),f=await Ht(p,t,l),h=Nt.get(p).js_modules?.main,d=et(Pt.get(a)),g=((e,t,n,r)=>({type:t,config:n,interpreter:r,io:Fe.get(r),run:(t,...n)=>e.run(r,t,...n),runAsync:(t,...n)=>e.runAsync(r,t,...n),runEvent:(...t)=>e.runEvent(r,...t)}))(d,i||a,l,f);let y="run";if(u&&(y+="Async"),r){const e=((e,t)=>(n,r)=>{const s=e[t].bind(e);e[t]=(e,t,...o)=>s(e,`${r?n:t}\\n${r?t:n}`,...o)})(d,y);let t,n,s="",o="";for(const e of Dt){const t=r[e];if(t){const n=e.endsWith("Async");(n&&u||!n&&!u)&&(e.startsWith("codeBefore")?s=t:o=t)}}o&&e(o,!1),s&&e(s,!0);for(const e of Lt.slice(2)){const s=r[e];if(s){const r=e.endsWith("Async");if(r&&u||!r&&!u){const r=ct(s);e.startsWith("onBefore")?t=r:n=r}}}Ut(d,g,Qt,u,t,n)}const{CustomEvent:m,document:w}=Vt,_=s&&w.getElementById(s)||null,b=e=>it(_,i||a,e,!0,m),E=wt(Vt,Xt,h);let k="";return((e,t,n,r)=>{if("pyodide"===e)return;const s="polyscript.js_modules";for(const e of Reflect.ownKeys(r))t.registerJSModule(n,`${s}.${e}`,ft(r,e));t.registerJSModule(n,s,r)})(a,d,f,E),d.registerJSModule(f,"polyscript",{xworker:Qt,currentScript:_,js_modules:E,get target(){return!k&&_&&("SCRIPT"===o?_.after(Qe(w.createElement(`script-${i||a}`),{id:k=`${s}-target`})):(k=s,_.replaceChildren(),_.style.display="block")),k}}),qt=d.runEvent.bind(d,f),Gt=d.transform.bind(d,f),_&&b("ready"),r?.onReady&&ct(r?.onReady).call(d,Bt.call(d,g,f),Qt),await d[y](f,n),_&&b("done"),f}catch(e){postMessage(e)}})(),Kt("error"),Kt("message"),Kt("messageerror")}));\n'],{type:"application/javascript"})),{type:"module"}),{postMessage:s}=r,o=this instanceof rn;if(e.length){const[t,r]=e;(n=qe({},n||{type:t,version:r})).type||(n.type=t)}const[a]=Kt(n.config),i=fetch(t).then(Fe).then((e=>{const t=o?this.toJSON():void 0;s.call(r,{options:n,config:a,code:e,hooks:t})})),c=qe(Ne(r,h).proxy,{importJS:ut,importCSS:pt});return Ke(r,{sync:{value:c},postMessage:{value:(e,...t)=>i.then((()=>s.call(r,e,...t)))},onerror:{writable:!0,configurable:!0,value:console.error}}),r.addEventListener("message",(e=>{const{data:t}=e;t instanceof Error&&(e.stopImmediatePropagation(),r.onerror(Ge(e,{type:{value:"error"},error:{value:t}})))})),o&&this.onWorker?.(this.interpreter,r),r};const on="Invalid content",an="Invalid worker attribute",cn="Invalid worker attribute";var ln=e=>{const{src:t,worker:n}=e.attributes;if(n){let{value:r}=n;if(r)throw new SyntaxError(cn);if(r=t?.value,!r){if(t)throw new SyntaxError(an);if(e.childElementCount){const{innerHTML:t,localName:n,type:s}=e,o=s||n.replace(/-script$/,"");r=Be(t),console.warn(`Deprecated: use <script type="${o}"> for an always safe content parsing:\n`,r)}else r=e.textContent;return URL.createObjectURL(new Blob([Je(r)],{type:"text/plain"}))}return r}if(t&&e.textContent.replace(/\/\*[\s\S]*?\*\//g,"").replace(/^\s*(?:\/\/|#).*/gm,"").trim())throw new SyntaxError(on)};const un=(e,t)=>{const n=(e=>{let t=e;for(;t.parentNode;)t=t.parentNode;return t})(e);return n.getElementById(t)||((e,t=document)=>t.querySelector(e))(t,n)},pn=new WeakMap,fn={get(){let e=pn.get(this);return e||(e=document.createElement(`${this.type}-script`),pn.set(this,e),mn(this)),e},set(e){"string"==typeof e?pn.set(this,un(this,e)):(pn.set(this,e),mn(this))}},hn=new WeakMap,dn=new Map,gn=(e,t)=>{const n=e?.value;return n?t+n:""},yn=(e,t,n,r,s,o=e)=>{if(!dn.has(t)){const a={interpreter:Yt(n,s),queue:Ve(),XWorker:sn(e,r)};dn.set(t,a),dn.has(e)||dn.set(e,a),dn.has(o)||dn.set(o,a)}return dn.get(t)},mn=async e=>{if(hn.has(e)){const{target:t}=e;t&&(e.closest("head")?document.body.append(t):e.after(t))}else{const{attributes:{async:t,config:n,env:r,target:s,version:o},src:a,type:i}=e,c=o?.value,l=zt(i,c);let u=gn(n,"|");const p=gn(r,"")||`${l}${u}`;u=u.slice(1);const f=ln(e);if(f){const n=new(sn(i,c))(f,{...et(e,i),async:!!t,config:u});return void hn.set(Xe(e,"xworker",{value:n}),{xworker:n})}const h=gn(s,""),d=yn(i,p,l,c,u);hn.set(Xe(e,"target",fn),d),h&&pn.set(e,un(e,h));const g=a?fetch(a).then(Fe):e.textContent;d.queue=d.queue.then((()=>(async(e,t,n,r)=>{const{type:s}=e,o=Lt.get(s);o.experimental&&console.warn(`The ${s} interpreter is experimental`);const[a,i]=await ze([hn.get(e).interpreter,t]);try{Xe(document,"currentScript",{configurable:!0,get:()=>e}),lt(s,o,a,at),o.registerJSModule(a,"polyscript",{XWorker:n,currentScript:e,js_modules:at}),tt(e,s,"ready");const t=o[r?"runAsync":"run"](a,i),c=tt.bind(null,e,s,"done");return r?t.then(c):c(),t}finally{delete document.currentScript}})(e,g,d.XWorker,!!t)))}},wn=new Proxy(Ge(null),{get:(e,t)=>new Promise(queueMicrotask).then((()=>_n(t)))}),_n=async e=>{if(dn.has(e)){const{interpreter:t,queue:n}=dn.get(e);return(await ze([t,n]))[0]}const t=dn.size?`Available interpreters are: ${[...dn.keys()].map((e=>`"${e}"`)).join(", ")}.`:"There are no interpreters in this page.";throw new Error(`The interpreter "${e}" was not found. ${t}`)},bn=async e=>{const{type:n,currentTarget:r}=e;if(Bt.length)for(let{name:s,value:o,ownerElement:a}of t(`./@*[${Bt.map((e=>`name()="${e}${n}"`)).join(" or ")}]`,r)){s=s.slice(0,-(n.length+1));const t=await _n(a.getAttribute(`${s}-env`)||s);Lt.get(s).runEvent(t,o,e)}},En=e=>{if(Bt.length)for(let{name:n,ownerElement:r}of t(`.//@*[${Bt.map((e=>`starts-with(name(),"${e}")`)).join(" or ")}]`,e)){const e=n.lastIndexOf("-"),t=n.slice(e+1);"env"!==t&&(r.addEventListener(t,bn),"disabled"in r&&!r.disabled&&(r.disabled=!0,wn[n.slice(0,e)].then((()=>{r.disabled=!1}))))}},vn=sn(),kn=[],Sn=new Map,xn=new Map,Tn=e=>{for(const t of kn)if(e.matches(t)){const n=Sn.get(t),r=An.get(n),{resolve:s}=xn.get(n),{options:o,known:a}=r;if(!a.has(e)){a.add(e);const{interpreter:t,config:i,version:c,env:l,onerror:u,hooks:p}=o;let f;try{const r=ln(e);if(r){const o=vn.call(new rn(null,p),r,{...et(e,n),version:c,type:t,custom:n,config:e.getAttribute("config")||i||{},async:e.hasAttribute("async")});return Xe(e,"xworker",{value:o}),void s({type:n,xworker:o})}}catch(e){f=e}const h=zt(t,c),d=l||`${h}${i?`|${i}`:""}`,{interpreter:g,XWorker:y}=yn(n,d,h,c,i,t);g.then((o=>{const a=Ge(Lt.get(t)),i=new rn(o,p),c=function(...e){return y.apply(i,e)},l={...nt(a,n,structuredClone(Dt.get(h)),o),XWorker:c};lt(t,a,o,at),a.registerJSModule(o,"polyscript",{XWorker:c,currentScript:n.startsWith("_")?null:e,js_modules:at});for(const t of["Run","RunAsync"]){const n=rt(a,`r${t.slice(1)}`);let r,s,o="",i="";for(const e of Qt){const n=p?.main?.[e];n&&e.endsWith(t)&&(e.startsWith("codeBefore")?o=Je(n()):i=Je(n()))}i&&n(i,!1),o&&n(o,!0);for(let e=2;e<en.length;e++){const n=en[e],o=p?.main?.[n];o&&n.endsWith(t)&&(n.startsWith("onBefore")?r=o:s=o)}nn(a,l,e,t.endsWith("Async"),r,s)}r.queue=r.queue.then((()=>(s(l),f&&u?.(f,e),p?.main?.onReady?.(l,e))))}))}}},An=new Map;let On=0;const jn=e=>(xn.has(e)||xn.set(e,Promise.withResolvers()),xn.get(e).promise),[{define:Rn,whenDefined:$n,env:Pn,Hook:Nn,XWorker:In},Mn]=((e,t,n=globalThis)=>{const r=Symbol.for(e),s=r in n;return[s?n[r]:Object.defineProperty(n,r,{value:t})[r],s]})("polyscript",{define:(t,n)=>{let r=null==t;if(r)t="_ps"+On++;else if(Lt.has(t)||An.has(t))throw new Error(`<script type="${t}"> already registered`);if(!Lt.has(n?.interpreter))throw new Error("Unspecified interpreter");Lt.set(t,Lt.get(n.interpreter));const s=[`script[type="${t}"]`];if(jn(t),r){const{hooks:e}=n,r=e?.main?.onReady;n={...n,hooks:{...e,main:{...e?.main,onReady(e,n){kn.splice(kn.indexOf(t),1),Lt.delete(t),An.delete(t),xn.delete(t),n.remove(),r?.(e)}}}},document.head.append(qe(document.createElement("script"),{type:t}))}else s.push(`${t}-script`),Bt.push(`${t}-`);for(const e of s)Sn.set(e,t);kn.push(...s),An.set(t,{options:qe({env:t},n),known:new WeakSet,queue:Promise.resolve()}),r||En(document),e(s.join(",")).forEach(Tn)},whenDefined:jn,env:wn,Hook:rn,XWorker:vn});if(!Mn){const t=new MutationObserver((e=>{const t=Jt.join(",");for(const{type:r,target:s,attributeName:o,addedNodes:a}of e)if("attributes"!==r)for(const e of a)1===e.nodeType&&(En(e),t&&e.matches(t)?mn(e):n(t,e,!0));else{const e=o.lastIndexOf("-")+1;if(e){const t=o.slice(0,e);for(const n of Bt)if(t===n){const t=o.slice(e);if("env"!==t){const e=s.hasAttribute(o)?"add":"remove";s[`${e}EventListener`](t,bn)}break}}}})),n=(t,n,r)=>{t&&e(t,n).forEach(mn),(t=kn.join(","))&&(r&&Tn(n),e(t,n).forEach(Tn))},r=e=>(t.observe(e,{childList:!0,subtree:!0,attributes:!0}),e),{attachShadow:s}=Element.prototype;qe(Element.prototype,{attachShadow(e){return r(s.call(this,e))}}),queueMicrotask((()=>{En(r(document)),n(Jt.join(","),document,!1)}))}export{Nn as Hook,on as INVALID_CONTENT,an as INVALID_SRC_ATTR,cn as INVALID_WORKER_ATTR,In as XWorker,Rn as define,Pn as env,$n as whenDefined};
+const e=(e,t=document)=>[...t.querySelectorAll(e)],t=(e,t=document)=>{const n=(new XPathEvaluator).createExpression(e).evaluate(t,XPathResult.ORDERED_NODE_SNAPSHOT_TYPE),r=[];for(let e=0,{snapshotLength:t}=n;e<t;e++)r.push(n.snapshotItem(e));return r},n="object"==typeof self?self:globalThis,r=e=>((e,t)=>{const r=(t,n)=>(e.set(n,t),t),s=o=>{if(e.has(o))return e.get(o);const[a,i]=t[o];switch(a){case 0:case-1:return r(i,o);case 1:{const e=r([],o);for(const t of i)e.push(s(t));return e}case 2:{const e=r({},o);for(const[t,n]of i)e[s(t)]=s(n);return e}case 3:return r(new Date(i),o);case 4:{const{source:e,flags:t}=i;return r(new RegExp(e,t),o)}case 5:{const e=r(new Map,o);for(const[t,n]of i)e.set(s(t),s(n));return e}case 6:{const e=r(new Set,o);for(const t of i)e.add(s(t));return e}case 7:{const{name:e,message:t}=i;return r(new n[e](t),o)}case 8:return r(BigInt(i),o);case"BigInt":return r(Object(BigInt(i)),o)}return r(new n[a](i),o)};return s})(new Map,e)(0),s="",{toString:o}={},{keys:a}=Object,i=e=>{const t=typeof e;if("object"!==t||!e)return[0,t];const n=o.call(e).slice(8,-1);switch(n){case"Array":return[1,s];case"Object":return[2,s];case"Date":return[3,s];case"RegExp":return[4,s];case"Map":return[5,s];case"Set":return[6,s]}return n.includes("Array")?[1,n]:n.includes("Error")?[7,n]:[2,n]},c=([e,t])=>0===e&&("function"===t||"symbol"===t),l=(e,{json:t,lossy:n}={})=>{const r=[];return((e,t,n,r)=>{const s=(e,t)=>{const s=r.push(e)-1;return n.set(t,s),s},o=r=>{if(n.has(r))return n.get(r);let[l,u]=i(r);switch(l){case 0:{let t=r;switch(u){case"bigint":l=8,t=r.toString();break;case"function":case"symbol":if(e)throw new TypeError("unable to serialize "+u);t=null;break;case"undefined":return s([-1],r)}return s([l,t],r)}case 1:{if(u)return s([u,[...r]],r);const e=[],t=s([l,e],r);for(const t of r)e.push(o(t));return t}case 2:{if(u)switch(u){case"BigInt":return s([u,r.toString()],r);case"Boolean":case"Number":case"String":return s([u,r.valueOf()],r)}if(t&&"toJSON"in r)return o(r.toJSON());const n=[],p=s([l,n],r);for(const t of a(r))!e&&c(i(r[t]))||n.push([o(t),o(r[t])]);return p}case 3:return s([l,r.toISOString()],r);case 4:{const{source:e,flags:t}=r;return s([l,{source:e,flags:t}],r)}case 5:{const t=[],n=s([l,t],r);for(const[n,s]of r)(e||!c(i(n))&&!c(i(s)))&&t.push([o(n),o(s)]);return n}case 6:{const t=[],n=s([l,t],r);for(const n of r)!e&&c(i(n))||t.push(o(n));return n}}const{message:p}=r;return s([l,{name:u,message:p}],r)};return o})(!(t||n),!!t,new Map,r)(e),r},{parse:u,stringify:p}=JSON,f={json:!0,lossy:!0};var h=Object.freeze({__proto__:null,parse:e=>r(u(e)),stringify:e=>p(l(e,f))});const d="ac6c1995-0675-46af-a99f-7fa298fa4e6d",g="M"+d,y="T"+d,m="array",w="function",_="null",b="number",E="object",v="string",k="symbol",S="undefined",x="apply",T="construct",A="defineProperty",O="deleteProperty",j="get",R="getOwnPropertyDescriptor",$="getPrototypeOf",P="has",I="isExtensible",N="ownKeys",M="preventExtensions",F="set",H="setPrototypeOf",W="delete";var L=e=>({value:new Promise((t=>{let n=new Worker("data:application/javascript,onmessage%3D(%7Bdata%3Ab%7D)%3D%3E(Atomics.wait(b%2C0)%2CpostMessage(0))");n.onmessage=t,n.postMessage(e)}))})
+/*! (c) Andrea Giammarchi - ISC */;const{Int32Array:C,Map:D,SharedArrayBuffer:U,Uint16Array:B}=globalThis,{BYTES_PER_ELEMENT:J}=C,{BYTES_PER_ELEMENT:q}=B,{isArray:G}=Array,{notify:K,wait:X,waitAsync:Y}=Atomics,z=new WeakSet,V=new WeakMap,Z={value:{then:e=>e()}};let Q=0;const ee=(e,{parse:t=JSON.parse,stringify:n=JSON.stringify,transform:r,interrupt:s}=JSON)=>{if(!V.has(e)){const o=(t,...n)=>e.postMessage({[d]:n},{transfer:t}),a=typeof s===w?s:s?.handler,i=s?.delay||42,c=new TextDecoder("utf-16"),l=(e,t)=>e?(Y||L)(t,0):(a?((e,t,n)=>{for(;"timed-out"===X(e,0,0,t);)n()})(t,i,a):X(t,0),Z);let u=!1;V.set(e,new Proxy(new D,{[P]:(e,t)=>"string"==typeof t&&!t.startsWith("_"),[j]:(n,s)=>"then"===s?null:(...n)=>{const a=Q++;let i=new C(new U(2*J)),p=[];z.has(n.at(-1)||p)&&z.delete(p=n.pop()),o(p,a,i,s,r?n.map(r):n);const f=e!==globalThis;let h=0;return u&&f&&(h=setTimeout(console.warn,1e3,`💀🔒 - Possible deadlock if proxy.${s}(...args) is awaited`)),l(f,i).value.then((()=>{clearTimeout(h);const e=i[1];if(!e)return;const n=q*e;return i=new C(new U(n+n%J)),o([],a,i),l(f,i).value.then((()=>t(c.decode(new B(i.buffer).slice(0,e)))))}))},[F](t,s,o){const a=typeof o;if(a!==w)throw new Error(`Unable to assign ${s} as ${a}`);if(!t.size){const s=new D;e.addEventListener("message",(async e=>{const o=e.data?.[d];if(G(o)){e.stopImmediatePropagation();const[a,i,...c]=o;let l;if(c.length){const[e,o]=c;if(t.has(e)){u=!0;try{const c=await t.get(e)(...o);if(void 0!==c){const e=n(r?r(c):c);s.set(a,e),i[1]=e.length}}catch(e){l=e}finally{u=!1}}else l=new Error(`Unsupported action: ${e}`);i[0]=1}else{const e=s.get(a);s.delete(a);for(let t=new B(i.buffer),n=0;n<e.length;n++)t[n]=e.charCodeAt(n)}if(K(i,0),l)throw l}}))}return!!t.set(s,o)}}))}return V.get(e)};ee.transfer=(...e)=>(z.add(e),e);const{isArray:te}=Array,ne=(e,t)=>t,re=e=>typeof e===w?(e=>e())(e):e;function se(){return this}const oe=(e,t)=>e===m?[t]:{t:e,v:t},ae=(e,t=ne)=>{let n=typeof e,r=e;return n===E&&(te(e)?(n=m,r=e.at(0)):({t:n,v:r}=e)),t(n,r)},ie=(e,t)=>e===w?t:oe(e,t),ce=(e,t=ie)=>{const n=null===e?_:typeof e;return t(n===E&&te(e)?m:n,e)},le=new FinalizationRegistry((([e,t,n])=>{n&&console.debug(`Held value ${String(t)} not relevant anymore`),e(t)})),ue=Object.create(null),pe=(e,t,{debug:n,return:r,token:s=e}=ue)=>{const o=r||new Proxy(e,ue),a=[o,[t,e,!!n]];return!1!==s&&a.push(s),le.register(...a),o},{defineProperty:fe,deleteProperty:he,getOwnPropertyDescriptor:de,getPrototypeOf:ge,isExtensible:ye,ownKeys:me,preventExtensions:we,set:_e,setPrototypeOf:be}=Reflect,{assign:Ee,create:ve}=Object,ke=ge(Int8Array),Se=(e,t)=>{const{get:n,set:r,value:s}=e;return n&&(e.get=t(n)),r&&(e.set=t(r)),s&&(e.value=t(s)),e},xe=e=>t=>ce(t,((t,n)=>{switch(t){case _:return oe(_,n);case E:if(n===globalThis)return oe(t,null);case m:case w:return e(t,n);case"boolean":case b:case v:case S:case"bigint":return oe(t,n);case k:{if(Te.has(n))return oe(t,Te.get(n));let e=Symbol.keyFor(n);if(e)return oe(t,`.${e}`)}}throw new TypeError(`Unable to handle this ${t}: ${String(n)}`)})),Te=new Map(me(Symbol).filter((e=>typeof Symbol[e]===k)).map((e=>[Symbol[e],e]))),Ae=e=>{if(e.startsWith("."))return Symbol.for(e.slice(1));for(const[t,n]of Te)if(n===e)return t},Oe=e=>e;var je=((e,t)=>{const n=t&&new WeakMap;if(t){const{addEventListener:e}=EventTarget.prototype;fe(EventTarget.prototype,"addEventListener",{value(t,r,...s){return s.at(0)?.invoke&&(n.has(this)||n.set(this,new Map),n.get(this).set(t,[].concat(s[0].invoke)),delete s[0].invoke),e.call(this,t,r,...s)}})}const r=t&&(e=>{const{currentTarget:t,target:r,type:s}=e;for(const o of n.get(t||r)?.get(s)||[])e[o]()});return function(n,s,o,...a){let i=0,c=this?.transform||Oe;const l=new Map,u=new Map,{[o]:p}=n,f=a.length?Ee(ve(globalThis),...a):globalThis,h=xe(((e,t)=>{if(!l.has(t)){let n;for(;u.has(n=i++););l.set(t,n),u.set(n,e===w?t:c(t))}return oe(e,l.get(t))})),d=e=>{p(W,oe(v,e))},g=(e,n)=>{switch(e){case E:if(null==n)return f;case m:if(typeof n===b)return u.get(n);if(!(n instanceof ke))for(const e in n)n[e]=y(n[e]);return n;case w:if(typeof n===v){if(!u.has(n)){const e=function(...e){return t&&e.at(0)instanceof Event&&r(...e),p(x,oe(w,n),h(this),e.map(h))};return u.set(n,new WeakRef(e)),pe(n,d,{return:e,token:!1})}return u.get(n).deref()}return u.get(n);case k:return Ae(n)}return n},y=e=>ae(e,g),_={[x]:(e,t,n)=>h(e.apply(t,n)),[T]:(e,t)=>h(new e(...t)),[A]:(e,t,n)=>h(fe(e,t,n)),[O]:(e,t)=>h(he(e,t)),[$]:e=>h(ge(e)),[j]:(e,t)=>h(e[t]),[R]:(e,t)=>{const n=de(e,t);return n?oe(E,Se(n,h)):oe(S,n)},[P]:(e,t)=>h(t in e),[I]:e=>h(ye(e)),[N]:e=>oe(m,me(e).map(h)),[M]:e=>h(we(e)),[F]:(e,t,n)=>h(_e(e,t,n)),[H]:(e,t)=>h(be(e,t)),[W](e){l.delete(u.get(e)),u.delete(e)}};return n[s]=(e,t,...n)=>{switch(e){case x:n[0]=y(n[0]),n[1]=n[1].map(y);break;case T:n[0]=n[0].map(y);break;case A:{const[e,t]=n;n[0]=y(e);const{get:r,set:s,value:o}=t;r&&(t.get=y(r)),s&&(t.set=y(s)),o&&(t.value=y(o));break}default:n=n.map(y)}return _[e](y(t),...n)},{proxy:n,[e.toLowerCase()]:f,[`is${e}Proxy`]:()=>!1}}})("Window",!0),Re=(e=>{let t=0;const n=new Map,r=new Map,s=Symbol();return function(o,a,i){const c=this?.transform||Oe,{[a]:l}=o,u=new Map,p=e=>{u.delete(e),l(W,f(e))},f=xe(((e,o)=>{if(s in o)return re(o[s]);if(e===w){if(o=c(o),!r.has(o)){let e;for(;r.has(e=String(t++)););n.set(o,e),r.set(e,o)}return oe(e,n.get(o))}if(!(o instanceof ke)){o=c(o);for(const e in o)o[e]=f(o[e])}return oe(e,o)})),h=(e,t,n)=>{if(!u.has(n)){const r=t===w?(e=>se.bind(e))(e):e,s=new Proxy(r,y);return u.set(n,new WeakRef(s)),pe(n,p,{return:s,token:!1})}return u.get(n).deref()},d=e=>ae(e,((t,n)=>{switch(t){case E:if(null===n)return globalThis;case m:return typeof n===b?h(e,t,n):n;case w:return typeof n===v?r.get(n):h(e,t,n);case k:return Ae(n)}return n})),g=(e,t,...n)=>d(l(e,re(t),...n)),y={[x]:(e,t,n)=>g(x,e,f(t),n.map(f)),[T]:(e,t)=>g(T,e,t.map(f)),[A]:(e,t,n)=>{const{get:r,set:s,value:o}=n;return typeof r===w&&(n.get=f(r)),typeof s===w&&(n.set=f(s)),typeof o===w&&(n.value=f(o)),g(A,e,f(t),n)},[O]:(e,t)=>g(O,e,f(t)),[$]:e=>g($,e),[j]:(e,t)=>t===s?e:g(j,e,f(t)),[R]:(e,t)=>{const n=g(R,e,f(t));return n&&Se(n,d)},[P]:(e,t)=>t===s||g(P,e,f(t)),[I]:e=>g(I,e),[N]:e=>g(N,e).map(d),[M]:e=>g(M,e),[F]:(e,t,n)=>g(F,e,f(t),f(n)),[H]:(e,t)=>g(H,e,f(t))};o[i]=(e,t,s,o)=>{switch(e){case x:return d(t).apply(d(s),o.map(d));case W:{const e=d(t);n.delete(r.get(e)),r.delete(e)}}};const _=new Proxy(oe(E,null),y);return{[e.toLowerCase()]:_,[`is${e}Proxy`]:e=>typeof e===E&&!!e&&s in e,proxy:o}}})("Window"),$e=typeof Worker===w?Worker:class{};const Pe=new WeakMap,Ie=(e,...t)=>{const n=ee(e,...t);if(!Pe.has(n)){const r=e instanceof $e?je:Re;Pe.set(n,r.call(t.at(0),n,g,y))}return Pe.get(n)};Ie.transfer=ee.transfer;Promise.withResolvers||(Promise.withResolvers=function(){var e,t,n=new this((function(n,r){e=n,t=r}));return{resolve:e,reject:t,promise:n}});const Ne=e=>e.arrayBuffer(),Me=e=>e.json(),Fe=e=>e.text();const He={object(...e){return this.string(function(e){for(var t=e[0],n=1,r=arguments.length;n<r;n++)t+=arguments[n]+e[n];return t}(...e))},string(e){for(const t of e.split(/[\r\n]+/))if(t.trim().length){/^(\s+)/.test(t)&&(e=e.replace(new RegExp("^"+RegExp.$1,"gm"),""));break}return e}},{replace:We}="",Le=/&(?:amp|#38|lt|#60|gt|#62|apos|#39|quot|#34);/g,Ce={"&amp;":"&","&#38;":"&","&lt;":"<","&#60;":"<","&gt;":">","&#62;":">","&apos;":"'","&#39;":"'","&quot;":'"',"&#34;":'"'},De=e=>Ce[e],Ue=new WeakMap,Be=e=>{const t=e||console,n={stderr:(t.stderr||console.error).bind(t),stdout:(t.stdout||console.log).bind(t)};return{stderr:(...e)=>n.stderr(...e),stdout:(...e)=>n.stdout(...e),async get(e){const t=await e;return Ue.set(t,n),t}}},Je=(e,...t)=>He[typeof e](e,...t),qe=e=>We.call(e,Le,De),{isArray:Ge}=Array,{assign:Ke,create:Xe,defineProperties:Ye,defineProperty:ze,entries:Ve}=Object,{all:Ze,resolve:Qe}=new Proxy(Promise,{get:(e,t)=>e[t].bind(e)}),et=(e,t=location.href)=>new URL(e,t.replace(/^blob:/,"")).href;let tt=0;const nt=(e,t)=>({id:e.id||(e.id=`${t}-w${tt++}`),tag:e.tagName}),rt=(e,t,n,r=!1,s=CustomEvent)=>{e.dispatchEvent(new s(`${t}:${n}`,{bubbles:!0,detail:{worker:r}}))},st=(e,t,n,r)=>({type:t,config:n,interpreter:r,io:Ue.get(r),run:(t,...n)=>e.run(r,t,...n),runAsync:(t,...n)=>e.runAsync(r,t,...n),runEvent:(...t)=>e.runEvent(r,...t)}),ot=(e,t)=>(n,r)=>{const s=e[t].bind(e);e[t]=(e,t,...o)=>s(e,`${r?n:t}\n${r?t:n}`,...o)},at=Symbol.for("polyscript.js_modules"),it=new Map;ze(globalThis,at,{value:it});const ct=new Proxy(it,{get:(e,t)=>e.get(t),has:(e,t)=>e.has(t),ownKeys:e=>[...e.keys()]}),lt=(e,t)=>!t.startsWith("_"),ut=(e,t)=>new Proxy(e,{has:lt,get:(e,n)=>e[t][n]}),pt=(e,t,n,r)=>{if("pyodide"===e)return;const s="polyscript.js_modules";for(const e of Reflect.ownKeys(r))t.registerJSModule(n,`${s}.${e}`,ut(r,e));t.registerJSModule(n,s,r)},ft=(e,t)=>import(e).then((e=>{it.set(t,{...e})})),ht=e=>new Promise(((t,n)=>{document.querySelector(`link[href="${e}"]`)&&t(),document.head.append(Ke(document.createElement("link"),{rel:"stylesheet",href:e,onload:t,onerror:n}))})),dt=e=>/\.css/i.test(new URL(e).pathname),gt=({FS:e,PATH:t,PATH_FS:n},r,s)=>{const o=n.resolve(r);return e.mkdirTree(t.dirname(o)),e.writeFile(o,new Uint8Array(s),{canOwn:!0})},yt=e=>{const t=e.split("/");return t.pop(),t.join("/")},mt=(e,t)=>{const n=[];for(const r of t.split("/"))"."!==r&&(n.push(r),r&&e.mkdir(n.join("/")))},wt=(e,t)=>{const n=[];for(const e of t.split("/"))switch(e){case"":case".":break;case"..":n.pop();break;default:n.push(e)}return[e.cwd()].concat(n).join("/").replace(/^\/+/,"/")},_t=e=>{const t=e.map((e=>e.trim().replace(/(^[/]*|[/]*$)/g,""))).filter((e=>""!==e&&"."!==e)).join("/");return e[0].startsWith("/")?`/${t}`:t},bt=(e,t)=>fetch(et(t,Et.get(e))),Et=new WeakMap,vt=(e,t,n)=>Ze((e=>{for(const{files:t,to_file:n,from:r=""}of e){if(void 0!==t&&void 0!==n)throw new Error("Cannot use 'to_file' and 'files' parameters together!");if(void 0===t&&void 0===n&&r.endsWith("/"))throw new Error(`Couldn't determine the filename from the path ${r}, please supply 'to_file' parameter.`)}return e.flatMap((({from:e="",to_folder:t=".",to_file:n,files:r})=>{if(Ge(r))return r.map((n=>({url:_t([e,n]),path:_t([t,n])})));const s=n||e.slice(1+e.lastIndexOf("/"));return[{url:e,path:_t([t,s])}]}))})(n).map((({url:r,path:s})=>bt(n,r).then(Ne).then((n=>e.writeFile(t,s,n)))))),kt=(e,t)=>t.endsWith("/")?`${t}${e.split("/").pop()}`:t,St=(e,t)=>e.replace(/\{.+?\}/g,(e=>{if(!t.has(e))throw new SyntaxError(`Invalid template: ${e}`);return t.get(e)})),xt=(e,t,n)=>Ze((e=>{const t=new Map,n=new Set,r=[];for(const[s,o]of Ve(e))if(/^\{.+\}$/.test(s)){if(t.has(s))throw new SyntaxError(`Duplicated template: ${s}`);t.set(s,St(o,t))}else{const e=St(s,t),a=kt(e,St(o||"./",t));if(n.has(a))throw new SyntaxError(`Duplicated destination: ${a}`);n.add(a),r.push({url:e,path:a})}return r})(n).map((({url:r,path:s})=>bt(n,r).then(Ne).then((n=>e.writeFile(t,s,n)))))),Tt=typeof document===S,At=({main:e,worker:t})=>{const n=[];if(t&&Tt)for(let[e,r]of Ve(t))e=et(e,Et.get(t)),n.push(ft(e,r));if(e&&!Tt)for(let[t,r]of Ve(e))t=et(t,Et.get(e)),dt(t)?ht(t):n.push(ft(t,r));return Ze(n)},Ot=(e,t,n)=>{e.registerJsModule(t,n)},jt=(e,t,...n)=>{try{return e.runPython(Je(t),...n)}catch(t){Ue.get(e).stderr(t)}},Rt=async(e,t,...n)=>{try{return await e.runPythonAsync(Je(t),...n)}catch(t){Ue.get(e).stderr(t)}},$t=async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.globals.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){Ue.get(e).stderr(t)}};var Pt=(new TextEncoder).encode('from uio import StringIO\nimport sys\n\nclass Response:\n    def __init__(self, f):\n        self.raw = f\n        self.encoding = "utf-8"\n        self._cached = None\n\n    def close(self):\n        if self.raw:\n            self.raw.close()\n            self.raw = None\n        self._cached = None\n\n    @property\n    def content(self):\n        if self._cached is None:\n            try:\n                self._cached = self.raw.read()\n            finally:\n                self.raw.close()\n                self.raw = None\n        return self._cached\n\n    @property\n    def text(self):\n        return str(self.content, self.encoding)\n\n    def json(self):\n        import ujson\n\n        return ujson.loads(self.content)\n\n\n# TODO try to support streaming xhr requests, a-la pyodide-http\nHEADERS_TO_IGNORE = ("user-agent",)\n\n\ntry:\n    import js\nexcept Exception as err:\n    raise OSError("This version of urequests can only be used in the browser")\n\n# TODO try to support streaming xhr requests, a-la pyodide-http\n\nHEADERS_TO_IGNORE = ("user-agent",)\n\n\ndef request(\n    method,\n    url,\n    data=None,\n    json=None,\n    headers={},\n    stream=None,\n    auth=None,\n    timeout=None,\n    parse_headers=True,\n):\n    from js import XMLHttpRequest\n\n    xhr = XMLHttpRequest.new()\n    xhr.withCredentials = False\n\n    if auth is not None:\n        import ubinascii\n\n        username, password = auth\n        xhr.open(method, url, False, username, password)\n    else:\n        xhr.open(method, url, False)\n\n    for name, value in headers.items():\n        if name.lower() not in HEADERS_TO_IGNORE:\n            xhr.setRequestHeader(name, value)\n\n    if timeout:\n        xhr.timeout = int(timeout * 1000)\n\n    if json is not None:\n        assert data is None\n        import ujson\n\n        data = ujson.dumps(json)\n        # s.write(b"Content-Type: application/json\\r\\n")\n        xhr.setRequestHeader("Content-Type", "application/json")\n\n    xhr.send(data)\n\n    # Emulates the construction process in the original urequests\n    resp = Response(StringIO(xhr.responseText))\n    resp.status_code = xhr.status\n    resp.reason = xhr.statusText\n    resp.headers = xhr.getAllResponseHeaders()\n\n    return resp\n\n\n# Other methods - head, post, put, patch, delete - are not used by\n# mip and therefore not included\n\n\ndef get(url, **kw):\n    return request("GET", url, **kw)\n\n\n# Content below this line is from the Micropython MIP package and is covered\n# by the applicable MIT license:\n# \n# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\n# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \n# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\n# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \n# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \n# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER \n# DEALINGS IN THE SOFTWARE.\n\n# MicroPython package installer\n# MIT license; Copyright (c) 2022 Jim Mussared\n\n\n_PACKAGE_INDEX = const("https://micropython.org/pi/v2")\n_CHUNK_SIZE = 128\n\n\n# This implements os.makedirs(os.dirname(path))\ndef _ensure_path_exists(path):\n    import os\n\n    split = path.split("/")\n\n    # Handle paths starting with "/".\n    if not split[0]:\n        split.pop(0)\n        split[0] = "/" + split[0]\n\n    prefix = ""\n    for i in range(len(split) - 1):\n        prefix += split[i]\n        try:\n            os.stat(prefix)\n        except:\n            os.mkdir(prefix)\n        prefix += "/"\n\n\n# Copy from src (stream) to dest (function-taking-bytes)\ndef _chunk(src, dest):\n    buf = memoryview(bytearray(_CHUNK_SIZE))\n    while True:\n        n = src.readinto(buf)\n        if n == 0:\n            break\n        dest(buf if n == _CHUNK_SIZE else buf[:n])\n\n\n# Check if the specified path exists and matches the hash.\ndef _check_exists(path, short_hash):\n    import os\n\n    try:\n        import binascii\n        import hashlib\n\n        with open(path, "rb") as f:\n            hs256 = hashlib.sha256()\n            _chunk(f, hs256.update)\n            existing_hash = str(binascii.hexlify(hs256.digest())[: len(short_hash)], "utf-8")\n            return existing_hash == short_hash\n    except:\n        return False\n\n\ndef _rewrite_url(url, branch=None):\n    if not branch:\n        branch = "HEAD"\n    if url.startswith("github:"):\n        url = url[7:].split("/")\n        url = (\n            "https://raw.githubusercontent.com/"\n            + url[0]\n            + "/"\n            + url[1]\n            + "/"\n            + branch\n            + "/"\n            + "/".join(url[2:])\n        )\n    return url\n\n\ndef _download_file(url, dest):\n    response = get(url)\n    try:\n        if response.status_code != 200:\n            print("Error", response.status_code, "requesting", url)\n            return False\n\n        print("Copying:", dest)\n        _ensure_path_exists(dest)\n        with open(dest, "wb") as f:\n            _chunk(response.raw, f.write)\n\n        return True\n    finally:\n        response.close()\n\n\ndef _install_json(package_json_url, index, target, version, mpy):\n    response = get(_rewrite_url(package_json_url, version))\n    try:\n        if response.status_code != 200:\n            print("Package not found:", package_json_url)\n            return False\n\n        package_json = response.json()\n    finally:\n        response.close()\n    for target_path, short_hash in package_json.get("hashes", ()):\n        fs_target_path = target + "/" + target_path\n        if _check_exists(fs_target_path, short_hash):\n            print("Exists:", fs_target_path)\n        else:\n            file_url = "{}/file/{}/{}".format(index, short_hash[:2], short_hash)\n            if not _download_file(file_url, fs_target_path):\n                print("File not found: {} {}".format(target_path, short_hash))\n                return False\n    for target_path, url in package_json.get("urls", ()):\n        fs_target_path = target + "/" + target_path\n        if not _download_file(_rewrite_url(url, version), fs_target_path):\n            print("File not found: {} {}".format(target_path, url))\n            return False\n    for dep, dep_version in package_json.get("deps", ()):\n        if not _install_package(dep, index, target, dep_version, mpy):\n            return False\n    return True\n\n\ndef _install_package(package, index, target, version, mpy):\n    if (\n        package.startswith("http://")\n        or package.startswith("https://")\n        or package.startswith("github:")\n    ):\n        if package.endswith(".py") or package.endswith(".mpy"):\n            print("Downloading {} to {}".format(package, target))\n            return _download_file(\n                _rewrite_url(package, version), target + "/" + package.rsplit("/")[-1]\n            )\n        else:\n            if not package.endswith(".json"):\n                if not package.endswith("/"):\n                    package += "/"\n                package += "package.json"\n            print("Installing {} to {}".format(package, target))\n    else:\n        if not version:\n            version = "latest"\n        print("Installing {} ({}) from {} to {}".format(package, version, index, target))\n\n        mpy_version = (\n            sys.implementation._mpy & 0xFF if mpy and hasattr(sys.implementation, "_mpy") else "py"\n        )\n\n        package = "{}/package/{}/{}/{}.json".format(index, mpy_version, package, version)\n\n    return _install_json(package, index, target, version, mpy)\n\n\ndef install(package, index=None, target=None, version=None, mpy=True):\n    if not target:\n        for p in sys.path:\n            if p.endswith("/lib"):\n                target = p\n                break\n        else:\n            print("Unable to find lib dir in sys.path")\n            return\n\n    if not index:\n        index = _PACKAGE_INDEX\n\n    if _install_package(package, index.rstrip("/"), target, version, mpy):\n        print("Done")\n    else:\n        print("Package may be partially installed")\n');var It={type:"micropython",module:(e="1.21.0-279")=>`https://cdn.jsdelivr.net/npm/@micropython/micropython-webassembly-pyscript@${e}/micropython.mjs`,async engine({loadMicroPython:e},t,n){const{stderr:r,stdout:s,get:o}=Be();n=n.replace(/\.m?js$/,".wasm");const a=await o(e({stderr:r,stdout:s,url:n}));if(t.files&&await xt(this,a,t.files),t.fetch&&await vt(this,a,t.fetch),t.js_modules&&await At(t.js_modules),this.writeFile(a,"./mip.py",Pt),t.packages){const e=a.pyimport("mip");for(const n of t.packages)e.install(n)}return a},registerJSModule:Ot,run:jt,runAsync:Rt,runEvent:$t,transform:(e,t)=>t instanceof e.PyProxy?e.PyProxy.toJs(t):t,writeFile:({FS:e,_module:{PATH:t,PATH_FS:n}},r,s)=>gt({FS:e,PATH:t,PATH_FS:n},r,s)};const Nt={dict_converter:Object.fromEntries};var Mt={type:"pyodide",module:(e="0.24.1")=>`https://cdn.jsdelivr.net/pyodide/v${e}/full/pyodide.mjs`,async engine({loadPyodide:e},t,n){const{stderr:r,stdout:s,get:o}=Be(),a=n.slice(0,n.lastIndexOf("/")),i=await o(e({stderr:r,stdout:s,indexURL:a}));if(t.files&&await xt(this,i,t.files),t.fetch&&await vt(this,i,t.fetch),t.js_modules&&await At(t.js_modules),t.packages){await i.loadPackage("micropip");const e=await i.pyimport("micropip");await e.install(t.packages,{keep_going:!0}),e.destroy()}return i},registerJSModule:Ot,run:jt,runAsync:Rt,runEvent:$t,transform:(e,t)=>t instanceof e.ffi.PyProxy?t.toJs(Nt):t,writeFile:({FS:e,PATH:t,_module:{PATH_FS:n}},r,s)=>gt({FS:e,PATH:t,PATH_FS:n},r,s)};const Ft="ruby-wasm-wasi",Ht=Ft.replace(/\W+/g,"_");var Wt={type:Ft,experimental:!0,module:(e="2.2.0")=>`https://cdn.jsdelivr.net/npm/ruby-3_2-wasm-wasi@${e}/dist/browser.esm.js`,async engine({DefaultRubyVM:e},t,n){const r=await fetch(`${n.slice(0,n.lastIndexOf("/"))}/ruby.wasm`),s=await WebAssembly.compile(await r.arrayBuffer()),{vm:o}=await e(s);return t.files&&await xt(this,o,t.files),t.fetch&&await vt(this,o,t.fetch),t.js_modules&&await At(t.js_modules),o},registerJSModule(e,t,n){t=t.replace(/\W+/g,"__");const r=`__module_${Ht}_${t}`;globalThis[r]=n,this.run(e,`require "js";$${t}=JS.global[:${r}]`),delete globalThis[r]},run:(e,t,...n)=>e.eval(Je(t),...n),runAsync:(e,t,...n)=>e.evalAsync(Je(t),...n),async runEvent(e,t,n){if(/^xworker\.(on\w+)$/.test(t)){const{$1:t}=RegExp,r=`__module_${Ht}_event`;globalThis[r]=n,this.run(e,`require "js";$xworker.call("${t}",JS.global[:${r}])`),delete globalThis[r]}else{const r=this.run(e,`method(:${t})`);await r.call(t,e.wrap(n))}},transform:(e,t)=>t,writeFile:()=>{throw new Error(`writeFile is not supported in ${Ft}`)}};var Lt={type:"wasmoon",module:(e="1.16.0")=>`https://cdn.jsdelivr.net/npm/wasmoon@${e}/+esm`,async engine({LuaFactory:e,LuaLibraries:t},n){const{stderr:r,stdout:s,get:o}=Be(),a=await o((new e).createEngine());return a.global.getTable(t.Base,(e=>{a.global.setField(e,"print",s),a.global.setField(e,"printErr",r)})),n.files&&await xt(this,a,n.files),n.fetch&&await vt(this,a,n.fetch),n.js_modules&&await At(n.js_modules),a},registerJSModule:(e,t,n)=>{e.global.set(t,n)},run:(e,t,...n)=>{try{return e.doStringSync(Je(t),...n)}catch(t){Ue.get(e).stderr(t)}},runAsync:async(e,t,...n)=>{try{return await e.doString(Je(t),...n)}catch(t){Ue.get(e).stderr(t)}},runEvent:async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.global.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){Ue.get(e).stderr(t)}},transform:(e,t)=>t,writeFile:({cmodule:{module:{FS:e}}},t,n)=>((e,t,n)=>(mt(e,yt(t)),t=wt(e,t),e.writeFile(t,new Uint8Array(n),{canOwn:!0})))(e,t,n)};const Ct=new Map,Dt=new Map,Ut=[],Bt=[],Jt=new Proxy(new Map,{get(e,t){if(!e.has(t)){const[n,...r]=t.split("@"),s=Ct.get(n),o=/^(?:\.?\.?\/|https?:\/\/)/i.test(r)?r.join("@"):s.module(...r);e.set(t,{url:o,module:import(o),engine:s.engine.bind(s)})}const{url:n,module:r,engine:s}=e.get(t);return(e,o)=>r.then((r=>{Dt.set(t,e);for(const t of["files","fetch"]){const n=e?.[t];n&&Et.set(n,o)}for(const t of["main","worker"]){const n=e?.js_modules?.[t];n&&Et.set(n,o)}return s(r,e,n)}))}}),qt=e=>{for(const t of[].concat(e.type))Ct.set(t,e),Ut.push(`script[type="${t}"]`),Bt.push(`${t}-`)};for(const e of[It,Mt,Wt,Lt])qt(e);const Gt=async e=>(await import("https://cdn.jsdelivr.net/npm/basic-toml@0.3.1/es.js")).parse(e),Kt=(e,t="./config.txt")=>{let n=typeof e;return"string"===n&&/\.(json|toml|txt)$/.test(e)?n=RegExp.$1:e=t,[et(e),n]},Xt=e=>{try{return JSON.parse(e)}catch(t){return Gt(e)}},Yt=(e,t,n,r={})=>{if(t){const[e,s]=Kt(t,n);"json"===s?r=fetch(e).then(Me):"toml"===s?r=fetch(e).then(Fe).then(Gt):"string"===s?r=Xt(t):"object"===s&&t?r=t:"txt"===s&&"string"==typeof r&&(r=Xt(r)),t=e}return Qe(r).then((n=>Jt[e](n,t)))},zt=(e,t="")=>`${e}@${t}`.replace(/@$/,"");function Vt(e=this){return String(e).replace(/^(async\s*)?(\bfunction\b)?(.*?)\(/,((e,t,n,r)=>r&&!n?`${t||""}function ${r}(`:e))}const Zt="BeforeRun",Qt="AfterRun",en=[`code${Zt}`,`code${Zt}Async`,`code${Qt}`,`code${Qt}Async`],tn=["onWorker","onReady",`on${Zt}`,`on${Zt}Async`,`on${Qt}`,`on${Qt}Async`];function nn(e,t){const{run:n,runAsync:r}=Ct.get(this.type);return{...e,run:n.bind(this,t),runAsync:r.bind(this,t)}}const rn=(e,t,n,r,s,o)=>{if(s||o){const a=nn.bind(e,t),i=r?"runAsync":"run",c=e[i];e[i]=r?async function(e,t,...r){s&&await s.call(this,a(e),n);const i=await c.call(this,e,t,...r);return o&&await o.call(this,a(e),n),i}:function(e,t,...r){s&&s.call(this,a(e),n);const i=c.call(this,e,t,...r);return o&&o.call(this,a(e),n),i}}};let sn=class{constructor(e,t={}){const{main:n,worker:r}=t;this.interpreter=e,this.onWorker=n?.onWorker;for(const e of tn.slice(1))this[e]=r?.[e];for(const e of en)this[e]=r?.[e]}toJSON(){const e={};for(const t of tn.slice(1))this[t]&&(e[t]=Vt(this[t]));for(const t of en)this[t]&&(e[t]=Je(this[t]()));return e}};var on=(...e)=>function(t,n){const r=new Worker(URL.createObjectURL(new Blob(['const e="object"==typeof self?self:globalThis,t=t=>((t,n)=>{const r=(e,n)=>(t.set(n,e),e),s=o=>{if(t.has(o))return t.get(o);const[a,i]=n[o];switch(a){case 0:case-1:return r(i,o);case 1:{const e=r([],o);for(const t of i)e.push(s(t));return e}case 2:{const e=r({},o);for(const[t,n]of i)e[s(t)]=s(n);return e}case 3:return r(new Date(i),o);case 4:{const{source:e,flags:t}=i;return r(new RegExp(e,t),o)}case 5:{const e=r(new Map,o);for(const[t,n]of i)e.set(s(t),s(n));return e}case 6:{const e=r(new Set,o);for(const t of i)e.add(s(t));return e}case 7:{const{name:t,message:n}=i;return r(new e[t](n),o)}case 8:return r(BigInt(i),o);case"BigInt":return r(Object(BigInt(i)),o)}return r(new e[a](i),o)};return s})(new Map,t)(0),n="",{toString:r}={},{keys:s}=Object,o=e=>{const t=typeof e;if("object"!==t||!e)return[0,t];const s=r.call(e).slice(8,-1);switch(s){case"Array":return[1,n];case"Object":return[2,n];case"Date":return[3,n];case"RegExp":return[4,n];case"Map":return[5,n];case"Set":return[6,n]}return s.includes("Array")?[1,s]:s.includes("Error")?[7,s]:[2,s]},a=([e,t])=>0===e&&("function"===t||"symbol"===t),i=(e,{json:t,lossy:n}={})=>{const r=[];return((e,t,n,r)=>{const i=(e,t)=>{const s=r.push(e)-1;return n.set(t,s),s},c=r=>{if(n.has(r))return n.get(r);let[l,u]=o(r);switch(l){case 0:{let t=r;switch(u){case"bigint":l=8,t=r.toString();break;case"function":case"symbol":if(e)throw new TypeError("unable to serialize "+u);t=null;break;case"undefined":return i([-1],r)}return i([l,t],r)}case 1:{if(u)return i([u,[...r]],r);const e=[],t=i([l,e],r);for(const t of r)e.push(c(t));return t}case 2:{if(u)switch(u){case"BigInt":return i([u,r.toString()],r);case"Boolean":case"Number":case"String":return i([u,r.valueOf()],r)}if(t&&"toJSON"in r)return c(r.toJSON());const n=[],p=i([l,n],r);for(const t of s(r))!e&&a(o(r[t]))||n.push([c(t),c(r[t])]);return p}case 3:return i([l,r.toISOString()],r);case 4:{const{source:e,flags:t}=r;return i([l,{source:e,flags:t}],r)}case 5:{const t=[],n=i([l,t],r);for(const[n,s]of r)(e||!a(o(n))&&!a(o(s)))&&t.push([c(n),c(s)]);return n}case 6:{const t=[],n=i([l,t],r);for(const n of r)!e&&a(o(n))||t.push(c(n));return n}}const{message:p}=r;return i([l,{name:u,message:p}],r)};return c})(!(t||n),!!t,new Map,r)(e),r},{parse:c,stringify:l}=JSON,u={json:!0,lossy:!0};var p=Object.freeze({__proto__:null,parse:e=>t(c(e)),stringify:e=>l(i(e,u))});const f="ac6c1995-0675-46af-a99f-7fa298fa4e6d",h="M"+f,d="T"+f,g="array",y="function",m="null",w="number",_="object",b="string",E="symbol",k="undefined",v="apply",T="construct",S="defineProperty",x="deleteProperty",A="get",O="getOwnPropertyDescriptor",j="getPrototypeOf",R="has",P="isExtensible",N="ownKeys",I="preventExtensions",$="set",M="setPrototypeOf",F="delete";var H=e=>({value:new Promise((t=>{let n=new Worker("data:application/javascript,onmessage%3D(%7Bdata%3Ab%7D)%3D%3E(Atomics.wait(b%2C0)%2CpostMessage(0))");n.onmessage=t,n.postMessage(e)}))})\n/*! (c) Andrea Giammarchi - ISC */;const{Int32Array:W,Map:C,SharedArrayBuffer:D,Uint16Array:L}=globalThis,{BYTES_PER_ELEMENT:B}=W,{BYTES_PER_ELEMENT:U}=L,{isArray:J}=Array,{notify:q,wait:G,waitAsync:K}=Atomics,Y=new WeakSet,z=new WeakMap,X={value:{then:e=>e()}};let V=0;const Z=(e,{parse:t=JSON.parse,stringify:n=JSON.stringify,transform:r,interrupt:s}=JSON)=>{if(!z.has(e)){const o=(t,...n)=>e.postMessage({[f]:n},{transfer:t}),a=typeof s===y?s:s?.handler,i=s?.delay||42,c=new TextDecoder("utf-16"),l=(e,t)=>e?(K||H)(t,0):(a?((e,t,n)=>{for(;"timed-out"===G(e,0,0,t);)n()})(t,i,a):G(t,0),X);let u=!1;z.set(e,new Proxy(new C,{[R]:(e,t)=>"string"==typeof t&&!t.startsWith("_"),[A]:(n,s)=>"then"===s?null:(...n)=>{const a=V++;let i=new W(new D(2*B)),p=[];Y.has(n.at(-1)||p)&&Y.delete(p=n.pop()),o(p,a,i,s,r?n.map(r):n);const f=e!==globalThis;let h=0;return u&&f&&(h=setTimeout(console.warn,1e3,`💀🔒 - Possible deadlock if proxy.${s}(...args) is awaited`)),l(f,i).value.then((()=>{clearTimeout(h);const e=i[1];if(!e)return;const n=U*e;return i=new W(new D(n+n%B)),o([],a,i),l(f,i).value.then((()=>t(c.decode(new L(i.buffer).slice(0,e)))))}))},[$](t,s,o){const a=typeof o;if(a!==y)throw new Error(`Unable to assign ${s} as ${a}`);if(!t.size){const s=new C;e.addEventListener("message",(async e=>{const o=e.data?.[f];if(J(o)){e.stopImmediatePropagation();const[a,i,...c]=o;let l;if(c.length){const[e,o]=c;if(t.has(e)){u=!0;try{const c=await t.get(e)(...o);if(void 0!==c){const e=n(r?r(c):c);s.set(a,e),i[1]=e.length}}catch(e){l=e}finally{u=!1}}else l=new Error(`Unsupported action: ${e}`);i[0]=1}else{const e=s.get(a);s.delete(a);for(let t=new L(i.buffer),n=0;n<e.length;n++)t[n]=e.charCodeAt(n)}if(q(i,0),l)throw l}}))}return!!t.set(s,o)}}))}return z.get(e)};Z.transfer=(...e)=>(Y.add(e),e);const{isArray:Q}=Array,ee=(e,t)=>t,te=e=>typeof e===y?(e=>e())(e):e;function ne(){return this}const re=(e,t)=>e===g?[t]:{t:e,v:t},se=(e,t=ee)=>{let n=typeof e,r=e;return n===_&&(Q(e)?(n=g,r=e.at(0)):({t:n,v:r}=e)),t(n,r)},oe=(e,t)=>e===y?t:re(e,t),ae=(e,t=oe)=>{const n=null===e?m:typeof e;return t(n===_&&Q(e)?g:n,e)},ie=new FinalizationRegistry((([e,t,n])=>{n&&console.debug(`Held value ${String(t)} not relevant anymore`),e(t)})),ce=Object.create(null),le=(e,t,{debug:n,return:r,token:s=e}=ce)=>{const o=r||new Proxy(e,ce),a=[o,[t,e,!!n]];return!1!==s&&a.push(s),ie.register(...a),o},{defineProperty:ue,deleteProperty:pe,getOwnPropertyDescriptor:fe,getPrototypeOf:he,isExtensible:de,ownKeys:ge,preventExtensions:ye,set:me,setPrototypeOf:we}=Reflect,{assign:_e,create:be}=Object,Ee=he(Int8Array),ke=(e,t)=>{const{get:n,set:r,value:s}=e;return n&&(e.get=t(n)),r&&(e.set=t(r)),s&&(e.value=t(s)),e},ve=e=>t=>ae(t,((t,n)=>{switch(t){case m:return re(m,n);case _:if(n===globalThis)return re(t,null);case g:case y:return e(t,n);case"boolean":case w:case b:case k:case"bigint":return re(t,n);case E:{if(Te.has(n))return re(t,Te.get(n));let e=Symbol.keyFor(n);if(e)return re(t,`.${e}`)}}throw new TypeError(`Unable to handle this ${t}: ${String(n)}`)})),Te=new Map(ge(Symbol).filter((e=>typeof Symbol[e]===E)).map((e=>[Symbol[e],e]))),Se=e=>{if(e.startsWith("."))return Symbol.for(e.slice(1));for(const[t,n]of Te)if(n===e)return t},xe=e=>e;var Ae=((e,t)=>{const n=t&&new WeakMap;if(t){const{addEventListener:e}=EventTarget.prototype;ue(EventTarget.prototype,"addEventListener",{value(t,r,...s){return s.at(0)?.invoke&&(n.has(this)||n.set(this,new Map),n.get(this).set(t,[].concat(s[0].invoke)),delete s[0].invoke),e.call(this,t,r,...s)}})}const r=t&&(e=>{const{currentTarget:t,target:r,type:s}=e;for(const o of n.get(t||r)?.get(s)||[])e[o]()});return function(n,s,o,...a){let i=0,c=this?.transform||xe;const l=new Map,u=new Map,{[o]:p}=n,f=a.length?_e(be(globalThis),...a):globalThis,h=ve(((e,t)=>{if(!l.has(t)){let n;for(;u.has(n=i++););l.set(t,n),u.set(n,e===y?t:c(t))}return re(e,l.get(t))})),d=e=>{p(F,re(b,e))},m=(e,n)=>{switch(e){case _:if(null==n)return f;case g:if(typeof n===w)return u.get(n);if(!(n instanceof Ee))for(const e in n)n[e]=H(n[e]);return n;case y:if(typeof n===b){if(!u.has(n)){const e=function(...e){return t&&e.at(0)instanceof Event&&r(...e),p(v,re(y,n),h(this),e.map(h))};return u.set(n,new WeakRef(e)),le(n,d,{return:e,token:!1})}return u.get(n).deref()}return u.get(n);case E:return Se(n)}return n},H=e=>se(e,m),W={[v]:(e,t,n)=>h(e.apply(t,n)),[T]:(e,t)=>h(new e(...t)),[S]:(e,t,n)=>h(ue(e,t,n)),[x]:(e,t)=>h(pe(e,t)),[j]:e=>h(he(e)),[A]:(e,t)=>h(e[t]),[O]:(e,t)=>{const n=fe(e,t);return n?re(_,ke(n,h)):re(k,n)},[R]:(e,t)=>h(t in e),[P]:e=>h(de(e)),[N]:e=>re(g,ge(e).map(h)),[I]:e=>h(ye(e)),[$]:(e,t,n)=>h(me(e,t,n)),[M]:(e,t)=>h(we(e,t)),[F](e){l.delete(u.get(e)),u.delete(e)}};return n[s]=(e,t,...n)=>{switch(e){case v:n[0]=H(n[0]),n[1]=n[1].map(H);break;case T:n[0]=n[0].map(H);break;case S:{const[e,t]=n;n[0]=H(e);const{get:r,set:s,value:o}=t;r&&(t.get=H(r)),s&&(t.set=H(s)),o&&(t.value=H(o));break}default:n=n.map(H)}return W[e](H(t),...n)},{proxy:n,[e.toLowerCase()]:f,[`is${e}Proxy`]:()=>!1}}})("Window",!0),Oe=(e=>{let t=0;const n=new Map,r=new Map,s=Symbol();return function(o,a,i){const c=this?.transform||xe,{[a]:l}=o,u=new Map,p=e=>{u.delete(e),l(F,f(e))},f=ve(((e,o)=>{if(s in o)return te(o[s]);if(e===y){if(o=c(o),!r.has(o)){let e;for(;r.has(e=String(t++)););n.set(o,e),r.set(e,o)}return re(e,n.get(o))}if(!(o instanceof Ee)){o=c(o);for(const e in o)o[e]=f(o[e])}return re(e,o)})),h=(e,t,n)=>{if(!u.has(n)){const r=t===y?(e=>ne.bind(e))(e):e,s=new Proxy(r,k);return u.set(n,new WeakRef(s)),le(n,p,{return:s,token:!1})}return u.get(n).deref()},d=e=>se(e,((t,n)=>{switch(t){case _:if(null===n)return globalThis;case g:return typeof n===w?h(e,t,n):n;case y:return typeof n===b?r.get(n):h(e,t,n);case E:return Se(n)}return n})),m=(e,t,...n)=>d(l(e,te(t),...n)),k={[v]:(e,t,n)=>m(v,e,f(t),n.map(f)),[T]:(e,t)=>m(T,e,t.map(f)),[S]:(e,t,n)=>{const{get:r,set:s,value:o}=n;return typeof r===y&&(n.get=f(r)),typeof s===y&&(n.set=f(s)),typeof o===y&&(n.value=f(o)),m(S,e,f(t),n)},[x]:(e,t)=>m(x,e,f(t)),[j]:e=>m(j,e),[A]:(e,t)=>t===s?e:m(A,e,f(t)),[O]:(e,t)=>{const n=m(O,e,f(t));return n&&ke(n,d)},[R]:(e,t)=>t===s||m(R,e,f(t)),[P]:e=>m(P,e),[N]:e=>m(N,e).map(d),[I]:e=>m(I,e),[$]:(e,t,n)=>m($,e,f(t),f(n)),[M]:(e,t)=>m(M,e,f(t))};o[i]=(e,t,s,o)=>{switch(e){case v:return d(t).apply(d(s),o.map(d));case F:{const e=d(t);n.delete(r.get(e)),r.delete(e)}}};const H=new Proxy(re(_,null),k);return{[e.toLowerCase()]:H,[`is${e}Proxy`]:e=>typeof e===_&&!!e&&s in e,proxy:o}}})("Window"),je=typeof Worker===y?Worker:class{};const Re=new WeakMap,Pe=(e,...t)=>{const n=Z(e,...t);if(!Re.has(n)){const r=e instanceof je?Ae:Oe;Re.set(n,r.call(t.at(0),n,h,d))}return Re.get(n)};Pe.transfer=Z.transfer;const Ne={object(...e){return this.string(function(e){for(var t=e[0],n=1,r=arguments.length;n<r;n++)t+=arguments[n]+e[n];return t}(...e))},string(e){for(const t of e.split(/[\\r\\n]+/))if(t.trim().length){/^(\\s+)/.test(t)&&(e=e.replace(new RegExp("^"+RegExp.$1,"gm"),""));break}return e}},Ie=new WeakMap,$e=e=>{const t=e||console,n={stderr:(t.stderr||console.error).bind(t),stdout:(t.stdout||console.log).bind(t)};return{stderr:(...e)=>n.stderr(...e),stdout:(...e)=>n.stdout(...e),async get(e){const t=await e;return Ie.set(t,n),t}}},Me=(e,...t)=>Ne[typeof e](e,...t),{isArray:Fe}=Array,{assign:He,create:We,defineProperties:Ce,defineProperty:De,entries:Le}=Object,{all:Be,resolve:Ue}=new Proxy(Promise,{get:(e,t)=>e[t].bind(e)}),Je=(e,t=location.href)=>new URL(e,t.replace(/^blob:/,"")).href,qe=(e,t,n,r=!1,s=CustomEvent)=>{e.dispatchEvent(new s(`${t}:${n}`,{bubbles:!0,detail:{worker:r}}))},Ge=e=>Function(`\'use strict\';return (${e})`)(),Ke=Symbol.for("polyscript.js_modules"),Ye=new Map;De(globalThis,Ke,{value:Ye}),new Proxy(Ye,{get:(e,t)=>e.get(t),has:(e,t)=>e.has(t),ownKeys:e=>[...e.keys()]});const ze=(e,t)=>!t.startsWith("_"),Xe=(e,t)=>new Proxy(e,{has:ze,get:(e,n)=>e[t][n]}),Ve=(e,t)=>import(e).then((e=>{Ye.set(t,{...e})})),Ze=e=>new Promise(((t,n)=>{document.querySelector(`link[href="${e}"]`)&&t(),document.head.append(He(document.createElement("link"),{rel:"stylesheet",href:e,onload:t,onerror:n}))})),Qe=e=>/\\.css/i.test(new URL(e).pathname);Promise.withResolvers||(Promise.withResolvers=function(){var e,t,n=new this((function(n,r){e=n,t=r}));return{resolve:e,reject:t,promise:n}});const et=e=>e.arrayBuffer(),tt=e=>e.json(),nt=e=>e.text(),rt=({FS:e,PATH:t,PATH_FS:n},r,s)=>{const o=n.resolve(r);return e.mkdirTree(t.dirname(o)),e.writeFile(o,new Uint8Array(s),{canOwn:!0})},st=e=>{const t=e.split("/");return t.pop(),t.join("/")},ot=(e,t)=>{const n=[];for(const r of t.split("/"))"."!==r&&(n.push(r),r&&e.mkdir(n.join("/")))},at=(e,t)=>{const n=[];for(const e of t.split("/"))switch(e){case"":case".":break;case"..":n.pop();break;default:n.push(e)}return[e.cwd()].concat(n).join("/").replace(/^\\/+/,"/")},it=e=>{const t=e.map((e=>e.trim().replace(/(^[/]*|[/]*$)/g,""))).filter((e=>""!==e&&"."!==e)).join("/");return e[0].startsWith("/")?`/${t}`:t},ct=(e,t)=>fetch(Je(t,lt.get(e))),lt=new WeakMap,ut=(e,t,n)=>Be((e=>{for(const{files:t,to_file:n,from:r=""}of e){if(void 0!==t&&void 0!==n)throw new Error("Cannot use \'to_file\' and \'files\' parameters together!");if(void 0===t&&void 0===n&&r.endsWith("/"))throw new Error(`Couldn\'t determine the filename from the path ${r}, please supply \'to_file\' parameter.`)}return e.flatMap((({from:e="",to_folder:t=".",to_file:n,files:r})=>{if(Fe(r))return r.map((n=>({url:it([e,n]),path:it([t,n])})));const s=n||e.slice(1+e.lastIndexOf("/"));return[{url:e,path:it([t,s])}]}))})(n).map((({url:r,path:s})=>ct(n,r).then(et).then((n=>e.writeFile(t,s,n)))))),pt=(e,t)=>t.endsWith("/")?`${t}${e.split("/").pop()}`:t,ft=(e,t)=>e.replace(/\\{.+?\\}/g,(e=>{if(!t.has(e))throw new SyntaxError(`Invalid template: ${e}`);return t.get(e)})),ht=(e,t,n)=>Be((e=>{const t=new Map,n=new Set,r=[];for(const[s,o]of Le(e))if(/^\\{.+\\}$/.test(s)){if(t.has(s))throw new SyntaxError(`Duplicated template: ${s}`);t.set(s,ft(o,t))}else{const e=ft(s,t),a=pt(e,ft(o||"./",t));if(n.has(a))throw new SyntaxError(`Duplicated destination: ${a}`);n.add(a),r.push({url:e,path:a})}return r})(n).map((({url:r,path:s})=>ct(n,r).then(et).then((n=>e.writeFile(t,s,n)))))),dt=typeof document===k,gt=({main:e,worker:t})=>{const n=[];if(t&&dt)for(let[e,r]of Le(t))e=Je(e,lt.get(t)),n.push(Ve(e,r));if(e&&!dt)for(let[t,r]of Le(e))t=Je(t,lt.get(e)),Qe(t)?Ze(t):n.push(Ve(t,r));return Be(n)},yt=(e,t)=>e.has(t),mt=e=>[...e.keys()];var wt=(e,t,n)=>{let r=globalThis[Ke],s="";if(n){s=lt.get(n);for(let[e,t]of Le(n)){let n=r.get(t);n&&!Fe(n)||(r.set(t,n||(n=[])),n.push(e))}}return((e,t,n,r)=>new Proxy(e,{has:yt,ownKeys:mt,get:(e,s)=>{let o=e.get(s);if(Fe(o)){let a=o;o=null;for(let e of a)e=Je(e,r),Qe(e)?n.importCSS(e):(n.importJS(e,s),o=t[Ke].get(s));e.set(s,o)}return o}}))(r,e,t,s)};const _t=(e,t,n)=>{e.registerJsModule(t,n)},bt=(e,t,...n)=>{try{return e.runPython(Me(t),...n)}catch(t){Ie.get(e).stderr(t)}},Et=async(e,t,...n)=>{try{return await e.runPythonAsync(Me(t),...n)}catch(t){Ie.get(e).stderr(t)}},kt=async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.globals.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){Ie.get(e).stderr(t)}};var vt=(new TextEncoder).encode(\'from uio import StringIO\\nimport sys\\n\\nclass Response:\\n    def __init__(self, f):\\n        self.raw = f\\n        self.encoding = "utf-8"\\n        self._cached = None\\n\\n    def close(self):\\n        if self.raw:\\n            self.raw.close()\\n            self.raw = None\\n        self._cached = None\\n\\n    @property\\n    def content(self):\\n        if self._cached is None:\\n            try:\\n                self._cached = self.raw.read()\\n            finally:\\n                self.raw.close()\\n                self.raw = None\\n        return self._cached\\n\\n    @property\\n    def text(self):\\n        return str(self.content, self.encoding)\\n\\n    def json(self):\\n        import ujson\\n\\n        return ujson.loads(self.content)\\n\\n\\n# TODO try to support streaming xhr requests, a-la pyodide-http\\nHEADERS_TO_IGNORE = ("user-agent",)\\n\\n\\ntry:\\n    import js\\nexcept Exception as err:\\n    raise OSError("This version of urequests can only be used in the browser")\\n\\n# TODO try to support streaming xhr requests, a-la pyodide-http\\n\\nHEADERS_TO_IGNORE = ("user-agent",)\\n\\n\\ndef request(\\n    method,\\n    url,\\n    data=None,\\n    json=None,\\n    headers={},\\n    stream=None,\\n    auth=None,\\n    timeout=None,\\n    parse_headers=True,\\n):\\n    from js import XMLHttpRequest\\n\\n    xhr = XMLHttpRequest.new()\\n    xhr.withCredentials = False\\n\\n    if auth is not None:\\n        import ubinascii\\n\\n        username, password = auth\\n        xhr.open(method, url, False, username, password)\\n    else:\\n        xhr.open(method, url, False)\\n\\n    for name, value in headers.items():\\n        if name.lower() not in HEADERS_TO_IGNORE:\\n            xhr.setRequestHeader(name, value)\\n\\n    if timeout:\\n        xhr.timeout = int(timeout * 1000)\\n\\n    if json is not None:\\n        assert data is None\\n        import ujson\\n\\n        data = ujson.dumps(json)\\n        # s.write(b"Content-Type: application/json\\\\r\\\\n")\\n        xhr.setRequestHeader("Content-Type", "application/json")\\n\\n    xhr.send(data)\\n\\n    # Emulates the construction process in the original urequests\\n    resp = Response(StringIO(xhr.responseText))\\n    resp.status_code = xhr.status\\n    resp.reason = xhr.statusText\\n    resp.headers = xhr.getAllResponseHeaders()\\n\\n    return resp\\n\\n\\n# Other methods - head, post, put, patch, delete - are not used by\\n# mip and therefore not included\\n\\n\\ndef get(url, **kw):\\n    return request("GET", url, **kw)\\n\\n\\n# Content below this line is from the Micropython MIP package and is covered\\n# by the applicable MIT license:\\n# \\n# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR\\n# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, \\n# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE\\n# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER \\n# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING \\n# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER \\n# DEALINGS IN THE SOFTWARE.\\n\\n# MicroPython package installer\\n# MIT license; Copyright (c) 2022 Jim Mussared\\n\\n\\n_PACKAGE_INDEX = const("https://micropython.org/pi/v2")\\n_CHUNK_SIZE = 128\\n\\n\\n# This implements os.makedirs(os.dirname(path))\\ndef _ensure_path_exists(path):\\n    import os\\n\\n    split = path.split("/")\\n\\n    # Handle paths starting with "/".\\n    if not split[0]:\\n        split.pop(0)\\n        split[0] = "/" + split[0]\\n\\n    prefix = ""\\n    for i in range(len(split) - 1):\\n        prefix += split[i]\\n        try:\\n            os.stat(prefix)\\n        except:\\n            os.mkdir(prefix)\\n        prefix += "/"\\n\\n\\n# Copy from src (stream) to dest (function-taking-bytes)\\ndef _chunk(src, dest):\\n    buf = memoryview(bytearray(_CHUNK_SIZE))\\n    while True:\\n        n = src.readinto(buf)\\n        if n == 0:\\n            break\\n        dest(buf if n == _CHUNK_SIZE else buf[:n])\\n\\n\\n# Check if the specified path exists and matches the hash.\\ndef _check_exists(path, short_hash):\\n    import os\\n\\n    try:\\n        import binascii\\n        import hashlib\\n\\n        with open(path, "rb") as f:\\n            hs256 = hashlib.sha256()\\n            _chunk(f, hs256.update)\\n            existing_hash = str(binascii.hexlify(hs256.digest())[: len(short_hash)], "utf-8")\\n            return existing_hash == short_hash\\n    except:\\n        return False\\n\\n\\ndef _rewrite_url(url, branch=None):\\n    if not branch:\\n        branch = "HEAD"\\n    if url.startswith("github:"):\\n        url = url[7:].split("/")\\n        url = (\\n            "https://raw.githubusercontent.com/"\\n            + url[0]\\n            + "/"\\n            + url[1]\\n            + "/"\\n            + branch\\n            + "/"\\n            + "/".join(url[2:])\\n        )\\n    return url\\n\\n\\ndef _download_file(url, dest):\\n    response = get(url)\\n    try:\\n        if response.status_code != 200:\\n            print("Error", response.status_code, "requesting", url)\\n            return False\\n\\n        print("Copying:", dest)\\n        _ensure_path_exists(dest)\\n        with open(dest, "wb") as f:\\n            _chunk(response.raw, f.write)\\n\\n        return True\\n    finally:\\n        response.close()\\n\\n\\ndef _install_json(package_json_url, index, target, version, mpy):\\n    response = get(_rewrite_url(package_json_url, version))\\n    try:\\n        if response.status_code != 200:\\n            print("Package not found:", package_json_url)\\n            return False\\n\\n        package_json = response.json()\\n    finally:\\n        response.close()\\n    for target_path, short_hash in package_json.get("hashes", ()):\\n        fs_target_path = target + "/" + target_path\\n        if _check_exists(fs_target_path, short_hash):\\n            print("Exists:", fs_target_path)\\n        else:\\n            file_url = "{}/file/{}/{}".format(index, short_hash[:2], short_hash)\\n            if not _download_file(file_url, fs_target_path):\\n                print("File not found: {} {}".format(target_path, short_hash))\\n                return False\\n    for target_path, url in package_json.get("urls", ()):\\n        fs_target_path = target + "/" + target_path\\n        if not _download_file(_rewrite_url(url, version), fs_target_path):\\n            print("File not found: {} {}".format(target_path, url))\\n            return False\\n    for dep, dep_version in package_json.get("deps", ()):\\n        if not _install_package(dep, index, target, dep_version, mpy):\\n            return False\\n    return True\\n\\n\\ndef _install_package(package, index, target, version, mpy):\\n    if (\\n        package.startswith("http://")\\n        or package.startswith("https://")\\n        or package.startswith("github:")\\n    ):\\n        if package.endswith(".py") or package.endswith(".mpy"):\\n            print("Downloading {} to {}".format(package, target))\\n            return _download_file(\\n                _rewrite_url(package, version), target + "/" + package.rsplit("/")[-1]\\n            )\\n        else:\\n            if not package.endswith(".json"):\\n                if not package.endswith("/"):\\n                    package += "/"\\n                package += "package.json"\\n            print("Installing {} to {}".format(package, target))\\n    else:\\n        if not version:\\n            version = "latest"\\n        print("Installing {} ({}) from {} to {}".format(package, version, index, target))\\n\\n        mpy_version = (\\n            sys.implementation._mpy & 0xFF if mpy and hasattr(sys.implementation, "_mpy") else "py"\\n        )\\n\\n        package = "{}/package/{}/{}/{}.json".format(index, mpy_version, package, version)\\n\\n    return _install_json(package, index, target, version, mpy)\\n\\n\\ndef install(package, index=None, target=None, version=None, mpy=True):\\n    if not target:\\n        for p in sys.path:\\n            if p.endswith("/lib"):\\n                target = p\\n                break\\n        else:\\n            print("Unable to find lib dir in sys.path")\\n            return\\n\\n    if not index:\\n        index = _PACKAGE_INDEX\\n\\n    if _install_package(package, index.rstrip("/"), target, version, mpy):\\n        print("Done")\\n    else:\\n        print("Package may be partially installed")\\n\');var Tt={type:"micropython",module:(e="1.21.0-279")=>`https://cdn.jsdelivr.net/npm/@micropython/micropython-webassembly-pyscript@${e}/micropython.mjs`,async engine({loadMicroPython:e},t,n){const{stderr:r,stdout:s,get:o}=$e();n=n.replace(/\\.m?js$/,".wasm");const a=await o(e({stderr:r,stdout:s,url:n}));if(t.files&&await ht(this,a,t.files),t.fetch&&await ut(this,a,t.fetch),t.js_modules&&await gt(t.js_modules),this.writeFile(a,"./mip.py",vt),t.packages){const e=a.pyimport("mip");for(const n of t.packages)e.install(n)}return a},registerJSModule:_t,run:bt,runAsync:Et,runEvent:kt,transform:(e,t)=>t instanceof e.PyProxy?e.PyProxy.toJs(t):t,writeFile:({FS:e,_module:{PATH:t,PATH_FS:n}},r,s)=>rt({FS:e,PATH:t,PATH_FS:n},r,s)};const St={dict_converter:Object.fromEntries};var xt={type:"pyodide",module:(e="0.24.1")=>`https://cdn.jsdelivr.net/pyodide/v${e}/full/pyodide.mjs`,async engine({loadPyodide:e},t,n){const{stderr:r,stdout:s,get:o}=$e(),a=n.slice(0,n.lastIndexOf("/")),i=await o(e({stderr:r,stdout:s,indexURL:a}));if(t.files&&await ht(this,i,t.files),t.fetch&&await ut(this,i,t.fetch),t.js_modules&&await gt(t.js_modules),t.packages){await i.loadPackage("micropip");const e=await i.pyimport("micropip");await e.install(t.packages,{keep_going:!0}),e.destroy()}return i},registerJSModule:_t,run:bt,runAsync:Et,runEvent:kt,transform:(e,t)=>t instanceof e.ffi.PyProxy?t.toJs(St):t,writeFile:({FS:e,PATH:t,_module:{PATH_FS:n}},r,s)=>rt({FS:e,PATH:t,PATH_FS:n},r,s)};const At="ruby-wasm-wasi",Ot=At.replace(/\\W+/g,"_");var jt={type:At,experimental:!0,module:(e="2.2.0")=>`https://cdn.jsdelivr.net/npm/ruby-3_2-wasm-wasi@${e}/dist/browser.esm.js`,async engine({DefaultRubyVM:e},t,n){const r=await fetch(`${n.slice(0,n.lastIndexOf("/"))}/ruby.wasm`),s=await WebAssembly.compile(await r.arrayBuffer()),{vm:o}=await e(s);return t.files&&await ht(this,o,t.files),t.fetch&&await ut(this,o,t.fetch),t.js_modules&&await gt(t.js_modules),o},registerJSModule(e,t,n){t=t.replace(/\\W+/g,"__");const r=`__module_${Ot}_${t}`;globalThis[r]=n,this.run(e,`require "js";$${t}=JS.global[:${r}]`),delete globalThis[r]},run:(e,t,...n)=>e.eval(Me(t),...n),runAsync:(e,t,...n)=>e.evalAsync(Me(t),...n),async runEvent(e,t,n){if(/^xworker\\.(on\\w+)$/.test(t)){const{$1:t}=RegExp,r=`__module_${Ot}_event`;globalThis[r]=n,this.run(e,`require "js";$xworker.call("${t}",JS.global[:${r}])`),delete globalThis[r]}else{const r=this.run(e,`method(:${t})`);await r.call(t,e.wrap(n))}},transform:(e,t)=>t,writeFile:()=>{throw new Error(`writeFile is not supported in ${At}`)}};var Rt={type:"wasmoon",module:(e="1.16.0")=>`https://cdn.jsdelivr.net/npm/wasmoon@${e}/+esm`,async engine({LuaFactory:e,LuaLibraries:t},n){const{stderr:r,stdout:s,get:o}=$e(),a=await o((new e).createEngine());return a.global.getTable(t.Base,(e=>{a.global.setField(e,"print",s),a.global.setField(e,"printErr",r)})),n.files&&await ht(this,a,n.files),n.fetch&&await ut(this,a,n.fetch),n.js_modules&&await gt(n.js_modules),a},registerJSModule:(e,t,n)=>{e.global.set(t,n)},run:(e,t,...n)=>{try{return e.doStringSync(Me(t),...n)}catch(t){Ie.get(e).stderr(t)}},runAsync:async(e,t,...n)=>{try{return await e.doString(Me(t),...n)}catch(t){Ie.get(e).stderr(t)}},runEvent:async(e,t,n)=>{const[r,...s]=t.split(".");let o,a=e.global.get(r);for(const e of s)[o,a]=[a,a[e]];try{await a.call(o,n)}catch(t){Ie.get(e).stderr(t)}},transform:(e,t)=>t,writeFile:({cmodule:{module:{FS:e}}},t,n)=>((e,t,n)=>(ot(e,st(t)),t=at(e,t),e.writeFile(t,new Uint8Array(n),{canOwn:!0})))(e,t,n)};const Pt=new Map,Nt=new Map,It=new Proxy(new Map,{get(e,t){if(!e.has(t)){const[n,...r]=t.split("@"),s=Pt.get(n),o=/^(?:\\.?\\.?\\/|https?:\\/\\/)/i.test(r)?r.join("@"):s.module(...r);e.set(t,{url:o,module:import(o),engine:s.engine.bind(s)})}const{url:n,module:r,engine:s}=e.get(t);return(e,o)=>r.then((r=>{Nt.set(t,e);for(const t of["files","fetch"]){const n=e?.[t];n&&lt.set(n,o)}for(const t of["main","worker"]){const n=e?.js_modules?.[t];n&&lt.set(n,o)}return s(r,e,n)}))}}),$t=e=>{for(const t of[].concat(e.type))Pt.set(t,e)};for(const e of[Tt,xt,jt,Rt])$t(e);const Mt=async e=>(await import("https://cdn.jsdelivr.net/npm/basic-toml@0.3.1/es.js")).parse(e),Ft=e=>{try{return JSON.parse(e)}catch(t){return Mt(e)}},Ht=(e,t,n,r={})=>{if(t){const[e,s]=((e,t="./config.txt")=>{let n=typeof e;return"string"===n&&/\\.(json|toml|txt)$/.test(e)?n=RegExp.$1:e=t,[Je(e),n]})(t,n);"json"===s?r=fetch(e).then(tt):"toml"===s?r=fetch(e).then(nt).then(Mt):"string"===s?r=Ft(t):"object"===s&&t?r=t:"txt"===s&&"string"==typeof r&&(r=Ft(r)),t=e}return Ue(r).then((n=>It[e](n,t)))},Wt="BeforeRun",Ct="AfterRun",Dt=[`code${Wt}`,`code${Wt}Async`,`code${Ct}`,`code${Ct}Async`],Lt=["onWorker","onReady",`on${Wt}`,`on${Wt}Async`,`on${Ct}`,`on${Ct}Async`];function Bt(e,t){const{run:n,runAsync:r}=Pt.get(this.type);return{...e,run:n.bind(this,t),runAsync:r.bind(this,t)}}const Ut=(e,t,n,r,s,o)=>{if(s||o){const a=Bt.bind(e,t),i=r?"runAsync":"run",c=e[i];e[i]=r?async function(e,t,...r){s&&await s.call(this,a(e),n);const i=await c.call(this,e,t,...r);return o&&await o.call(this,a(e),n),i}:function(e,t,...r){s&&s.call(this,a(e),n);const i=c.call(this,e,t,...r);return o&&o.call(this,a(e),n),i}}};try{new SharedArrayBuffer(4)}catch(e){throw new Error(["Unable to use SharedArrayBuffer due insecure environment.","Please read requirements in MDN: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer#security_requirements"].join("\\n"))}let Jt,qt,Gt;const Kt=(e,t)=>{addEventListener(e,t||(async t=>{try{await Jt,qt(`xworker.on${e}`,t)}catch(e){postMessage(e)}}),!!t&&{once:!0})},{parse:Yt,stringify:zt}=p,{proxy:Xt,window:Vt,isWindowProxy:Zt}=Pe(self,{parse:Yt,stringify:zt,transform:e=>Gt?Gt(e):e}),Qt={sync:Xt,window:Vt,isWindowProxy:Zt,onmessage:console.info,onerror:console.error,onmessageerror:console.warn,postMessage:postMessage.bind(self)};Kt("message",(({data:{options:e,config:t,configURL:n,code:r,hooks:s}})=>{Jt=(async()=>{try{const{id:o,tag:a,type:i,custom:c,version:l,config:u,async:p}=e,f=((e,t="")=>`${e}@${t}`.replace(/@$/,""))(i,l),h=await Ht(f,t,n,u),d=Nt.get(f).js_modules?.main,g=We(Pt.get(i)),y=((e,t,n,r)=>({type:t,config:n,interpreter:r,io:Ie.get(r),run:(t,...n)=>e.run(r,t,...n),runAsync:(t,...n)=>e.runAsync(r,t,...n),runEvent:(...t)=>e.runEvent(r,...t)}))(g,c||i,u,h);let m="run";if(p&&(m+="Async"),s){const e=((e,t)=>(n,r)=>{const s=e[t].bind(e);e[t]=(e,t,...o)=>s(e,`${r?n:t}\\n${r?t:n}`,...o)})(g,m);let t,n,r="",o="";for(const e of Dt){const t=s[e];if(t){const n=e.endsWith("Async");(n&&p||!n&&!p)&&(e.startsWith("codeBefore")?r=t:o=t)}}o&&e(o,!1),r&&e(r,!0);for(const e of Lt.slice(2)){const r=s[e];if(r){const s=e.endsWith("Async");if(s&&p||!s&&!p){const s=Ge(r);e.startsWith("onBefore")?t=s:n=s}}}Ut(g,y,Qt,p,t,n)}const{CustomEvent:w,document:_}=Vt,b=o&&_.getElementById(o)||null,E=e=>qe(b,c||i,e,!0,w),k=wt(Vt,Xt,d);let v="";return((e,t,n,r)=>{if("pyodide"===e)return;const s="polyscript.js_modules";for(const e of Reflect.ownKeys(r))t.registerJSModule(n,`${s}.${e}`,Xe(r,e));t.registerJSModule(n,s,r)})(i,g,h,k),g.registerJSModule(h,"polyscript",{xworker:Qt,currentScript:b,js_modules:k,get target(){return!v&&b&&("SCRIPT"===a?b.after(He(_.createElement(`script-${c||i}`),{id:v=`${o}-target`})):(v=o,b.replaceChildren(),b.style.display="block")),v}}),qt=g.runEvent.bind(g,h),Gt=g.transform.bind(g,h),b&&E("ready"),s?.onReady&&Ge(s?.onReady).call(g,Bt.call(g,y,h),Qt),await g[m](h,r),b&&E("done"),h}catch(e){postMessage(e)}})(),Kt("error"),Kt("message"),Kt("messageerror")}));\n'],{type:"application/javascript"})),{type:"module"}),{postMessage:s}=r,o=this instanceof sn;if(e.length){const[t,r]=e;(n=Ke({},n||{type:t,version:r})).type||(n.type=t)}const[a]=Kt(n.config,n.configURL),i=fetch(t).then(Fe).then((e=>{const t=o?this.toJSON():void 0;s.call(r,{options:n,config:a,code:e,hooks:t})})),c=Ke(Ie(r,h).proxy,{importJS:ft,importCSS:ht});return Ye(r,{sync:{value:c},postMessage:{value:(e,...t)=>i.then((()=>s.call(r,e,...t)))},onerror:{writable:!0,configurable:!0,value:console.error}}),r.addEventListener("message",(e=>{const{data:t}=e;t instanceof Error&&(e.stopImmediatePropagation(),r.onerror(Xe(e,{type:{value:"error"},error:{value:t}})))})),o&&this.onWorker?.(this.interpreter,r),r};const an="Invalid content",cn="Invalid worker attribute",ln="Invalid worker attribute";var un=e=>{const{src:t,worker:n}=e.attributes;if(n){let{value:r}=n;if(r)throw new SyntaxError(ln);if(r=t?.value,!r){if(t)throw new SyntaxError(cn);if(e.childElementCount){const{innerHTML:t,localName:n,type:s}=e,o=s||n.replace(/-script$/,"");r=qe(t),console.warn(`Deprecated: use <script type="${o}"> for an always safe content parsing:\n`,r)}else r=e.textContent;return URL.createObjectURL(new Blob([Je(r)],{type:"text/plain"}))}return r}if(t&&e.textContent.replace(/\/\*[\s\S]*?\*\//g,"").replace(/^\s*(?:\/\/|#).*/gm,"").trim())throw new SyntaxError(an)};const pn=(e,t)=>{const n=(e=>{let t=e;for(;t.parentNode;)t=t.parentNode;return t})(e);return n.getElementById(t)||((e,t=document)=>t.querySelector(e))(t,n)},fn=new WeakMap,hn={get(){let e=fn.get(this);return e||(e=document.createElement(`${this.type}-script`),fn.set(this,e),wn(this)),e},set(e){"string"==typeof e?fn.set(this,pn(this,e)):(fn.set(this,e),wn(this))}},dn=new WeakMap,gn=new Map,yn=(e,t)=>{const n=e?.value;return n?t+n:""},mn=(e,t,n,r,s,o,a=e)=>{if(!gn.has(t)){const i={interpreter:Yt(n,s,o),queue:Qe(),XWorker:on(e,r)};gn.set(t,i),gn.has(e)||gn.set(e,i),gn.has(a)||gn.set(a,i)}return gn.get(t)},wn=async e=>{if(dn.has(e)){const{target:t}=e;t&&(e.closest("head")?document.body.append(t):e.after(t))}else{const{attributes:{async:t,config:n,env:r,target:s,version:o},src:a,type:i}=e,c=o?.value,l=zt(i,c);let u=yn(n,"|");const p=yn(r,"")||`${l}${u}`;u=u.slice(1);const f=un(e);if(f){const n=new(on(i,c))(f,{...nt(e,i),async:!!t,config:u});return void dn.set(ze(e,"xworker",{value:n}),{xworker:n})}const h=yn(s,""),d=mn(i,p,l,c,u);dn.set(ze(e,"target",hn),d),h&&fn.set(e,pn(e,h));const g=a?fetch(a).then(Fe):e.textContent;d.queue=d.queue.then((()=>(async(e,t,n,r)=>{const{type:s}=e,o=Ct.get(s);o.experimental&&console.warn(`The ${s} interpreter is experimental`);const[a,i]=await Ze([dn.get(e).interpreter,t]);try{ze(document,"currentScript",{configurable:!0,get:()=>e}),pt(s,o,a,ct),o.registerJSModule(a,"polyscript",{XWorker:n,currentScript:e,js_modules:ct}),rt(e,s,"ready");const t=o[r?"runAsync":"run"](a,i),c=rt.bind(null,e,s,"done");return r?t.then(c):c(),t}finally{delete document.currentScript}})(e,g,d.XWorker,!!t)))}},_n=new Proxy(Xe(null),{get:(e,t)=>new Promise(queueMicrotask).then((()=>bn(t)))}),bn=async e=>{if(gn.has(e)){const{interpreter:t,queue:n}=gn.get(e);return(await Ze([t,n]))[0]}const t=gn.size?`Available interpreters are: ${[...gn.keys()].map((e=>`"${e}"`)).join(", ")}.`:"There are no interpreters in this page.";throw new Error(`The interpreter "${e}" was not found. ${t}`)},En=async e=>{const{type:n,currentTarget:r}=e;if(Bt.length)for(let{name:s,value:o,ownerElement:a}of t(`./@*[${Bt.map((e=>`name()="${e}${n}"`)).join(" or ")}]`,r)){s=s.slice(0,-(n.length+1));const t=await bn(a.getAttribute(`${s}-env`)||s);Ct.get(s).runEvent(t,o,e)}},vn=e=>{if(Bt.length)for(let{name:n,ownerElement:r}of t(`.//@*[${Bt.map((e=>`starts-with(name(),"${e}")`)).join(" or ")}]`,e)){const e=n.lastIndexOf("-"),t=n.slice(e+1);"env"!==t&&(r.addEventListener(t,En),"disabled"in r&&!r.disabled&&(r.disabled=!0,_n[n.slice(0,e)].then((()=>{r.disabled=!1}))))}},kn=on(),Sn=[],xn=new Map,Tn=new Map,An=e=>{for(const t of Sn)if(e.matches(t)){const n=xn.get(t),r=On.get(n),{resolve:s}=Tn.get(n),{options:o,known:a}=r;if(!a.has(e)){a.add(e);const{interpreter:t,configURL:i,config:c,version:l,env:u,onerror:p,hooks:f}=o;let h;try{const r=un(e);if(r){const o=kn.call(new sn(null,f),r,{...nt(e,n),version:l,configURL:i,type:t,custom:n,config:e.getAttribute("config")||c||{},async:e.hasAttribute("async")});return ze(e,"xworker",{value:o}),void s({type:n,xworker:o})}}catch(e){h=e}const d=zt(t,l),g=u||`${d}${c?`|${c}`:""}`,{interpreter:y,XWorker:m}=mn(n,g,d,l,c,i,t);y.then((o=>{const a=Xe(Ct.get(t)),i=new sn(o,f),c=function(...e){return m.apply(i,e)},l={...st(a,n,structuredClone(Dt.get(d)),o),XWorker:c};pt(t,a,o,ct),a.registerJSModule(o,"polyscript",{XWorker:c,currentScript:n.startsWith("_")?null:e,js_modules:ct});for(const t of["Run","RunAsync"]){const n=ot(a,`r${t.slice(1)}`);let r,s,o="",i="";for(const e of en){const n=f?.main?.[e];n&&e.endsWith(t)&&(e.startsWith("codeBefore")?o=Je(n()):i=Je(n()))}i&&n(i,!1),o&&n(o,!0);for(let e=2;e<tn.length;e++){const n=tn[e],o=f?.main?.[n];o&&n.endsWith(t)&&(n.startsWith("onBefore")?r=o:s=o)}rn(a,l,e,t.endsWith("Async"),r,s)}r.queue=r.queue.then((()=>(s(l),h&&p?.(h,e),f?.main?.onReady?.(l,e))))}))}}},On=new Map;let jn=0;const Rn=e=>(Tn.has(e)||Tn.set(e,Promise.withResolvers()),Tn.get(e).promise),[{define:$n,whenDefined:Pn,env:In,Hook:Nn,XWorker:Mn},Fn]=((e,t,n=globalThis)=>{const r=Symbol.for(e),s=r in n;return[s?n[r]:Object.defineProperty(n,r,{value:t})[r],s]})("polyscript",{define:(t,n)=>{let r=null==t;if(r)t="_ps"+jn++;else if(Ct.has(t)||On.has(t))throw new Error(`<script type="${t}"> already registered`);if(!Ct.has(n?.interpreter))throw new Error("Unspecified interpreter");Ct.set(t,Ct.get(n.interpreter));const s=[`script[type="${t}"]`];if(Rn(t),r){const{hooks:e}=n,r=e?.main?.onReady;n={...n,hooks:{...e,main:{...e?.main,onReady(e,n){Sn.splice(Sn.indexOf(t),1),Ct.delete(t),On.delete(t),Tn.delete(t),n.remove(),r?.(e)}}}},document.head.append(Ke(document.createElement("script"),{type:t}))}else s.push(`${t}-script`),Bt.push(`${t}-`);for(const e of s)xn.set(e,t);Sn.push(...s),On.set(t,{options:Ke({env:t},n),known:new WeakSet,queue:Promise.resolve()}),r||vn(document),e(s.join(",")).forEach(An)},whenDefined:Rn,env:_n,Hook:sn,XWorker:kn});if(!Fn){const t=new MutationObserver((e=>{const t=Ut.join(",");for(const{type:r,target:s,attributeName:o,addedNodes:a}of e)if("attributes"!==r)for(const e of a)1===e.nodeType&&(vn(e),t&&e.matches(t)?wn(e):n(t,e,!0));else{const e=o.lastIndexOf("-")+1;if(e){const t=o.slice(0,e);for(const n of Bt)if(t===n){const t=o.slice(e);if("env"!==t){const e=s.hasAttribute(o)?"add":"remove";s[`${e}EventListener`](t,En)}break}}}})),n=(t,n,r)=>{t&&e(t,n).forEach(wn),(t=Sn.join(","))&&(r&&An(n),e(t,n).forEach(An))},r=e=>(t.observe(e,{childList:!0,subtree:!0,attributes:!0}),e),{attachShadow:s}=Element.prototype;Ke(Element.prototype,{attachShadow(e){return r(s.call(this,e))}}),queueMicrotask((()=>{vn(r(document)),n(Ut.join(","),document,!1)}))}export{Nn as Hook,an as INVALID_CONTENT,cn as INVALID_SRC_ATTR,ln as INVALID_WORKER_ATTR,Mn as XWorker,$n as define,In as env,Pn as whenDefined};
 //# sourceMappingURL=core.js.map
