+++ esm/exports.js
@@ -1,5 +1,8 @@
 // this file simply exports enough stuff to allow
 // 3rd party libraries, including PyScript, to work
+import toJSONCallback from 'to-json-callback';
+export { toJSONCallback };
+
 export * from './fetch-utils.js';
 export * from './index.js';
 export * from './script-handler.js';
