+++ esm/interpreter/_io.js
@@ -1,0 +1,20 @@
+// REQUIRES INTEGRATION TEST
+/* c8 ignore start */
+export const io = new WeakMap();
+export const stdio = (init) => {
+    const context = init || console;
+    const localIO = {
+        stderr: (context.stderr || console.error).bind(context),
+        stdout: (context.stdout || console.log).bind(context),
+    };
+    return {
+        stderr: (...args) => localIO.stderr(...args),
+        stdout: (...args) => localIO.stdout(...args),
+        async get(engine) {
+            const interpreter = await engine;
+            io.set(interpreter, localIO);
+            return interpreter;
+        },
+    };
+};
+/* c8 ignore stop */
