+++ esm/interpreter/_utils.js
@@ -6,23 +6,6 @@ import { absoluteURL, all, entries, importCSS, importJS, isArray, isCSS } from '
 
 // REQUIRES INTEGRATION TEST
 /* c8 ignore start */
-export const io = new WeakMap();
-export const stdio = (init) => {
-    const context = init || console;
-    const localIO = {
-        stderr: (context.stderr || console.error).bind(context),
-        stdout: (context.stdout || console.log).bind(context),
-    };
-    return {
-        stderr: (...args) => localIO.stderr(...args),
-        stdout: (...args) => localIO.stdout(...args),
-        async get(engine) {
-            const interpreter = await engine;
-            io.set(interpreter, localIO);
-            return interpreter;
-        },
-    };
-};
 
 // This should be the only helper needed for all Emscripten based FS exports
 export const writeFile = ({ FS, PATH, PATH_FS }, path, buffer) => {
