+++ test/config/assets/a.py
@@ -1,0 +1,1 @@
+print("a")
