+++ test/config/index.py
@@ -1,0 +1,3 @@
+import a
+
+print("OK")
