+++ test/config/invent/pubsub.py
@@ -1,0 +1,1 @@
+print("pubsub")
