+++ test/config/main.py
@@ -1,0 +1,3 @@
+from invent import pubsub
+
+print("main OK")
