+++ test/config/src/invent/pubsub.py
@@ -1,0 +1,1 @@
+print("pubsub")
