+++ test/config/worker.py
@@ -1,0 +1,3 @@
+from invent import pubsub
+
+print("worker OK")
